### 2022-10-09 #4831 @editorialbot
[PRE REVIEW]: GraphNeT: Graph neural networks for neutrino telescope event reconstruction
**Submitting author:** @asogaard (<a href="http://orcid.org/0000-0002-0823-056X">Andreas Søgaard</a>)
**Repository:** https://github.com/graphnet-team/graphnet
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
