+++ src/graphnet/data/extractors/i3featureextractor.py
@@ -62,7 +62,9 @@ class I3FeatureExtractorIceCube86(I3FeatureExtractor):
             x = self._gcd_dict[om_key].position.x
             y = self._gcd_dict[om_key].position.y
             z = self._gcd_dict[om_key].position.z
-            event_time = frame["I3EventHeader"].start_time.mod_julian_day_double
+            event_time = frame[
+                "I3EventHeader"
+            ].start_time.mod_julian_day_double
             area = self._gcd_dict[om_key].area
             rde = self._get_relative_dom_efficiency(frame, om_key)
 
@@ -104,7 +106,7 @@ class I3FeatureExtractorIceCube86(I3FeatureExtractor):
                 output["is_saturated_dom"].append(is_saturated_dom)
                 output["is_errata_dom"].append(is_errata_dom)
                 output["event_time"].append(event_time)
-                
+
         return output
 
     def _get_relative_dom_efficiency(self, frame, om_key):
