### 2022-10-11 2098184
Changed event time extration to above for loop
