### 2022-10-11 3e50af2
Update tests/ml/test_loss_functions.py
PEP8-compliant function name

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
