### 2022-10-11 a133dfe
Update tests/ml/test_loss_functions.py
modified assertion

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
