### 2022-10-11 85a7695
Update tests/ml/test_loss_functions.py
Expanded combinations

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
