+++ src/graphnet/models/task/reconstruction.py
@@ -39,6 +39,22 @@ class AzimuthReconstruction(AzimuthReconstructionWithKappa):
         return angle
 
 
+class DirectionReconstructionWithKappa(Task):
+    """Reconstructs Direction with kappa from The 3DVonMisesFisher Distribution."""
+
+    # Requires 4 features: untransformed points in (x,y,z)-space + kappa estimation.
+    nb_inputs = 4
+
+    def _forward(self, x):
+        # Transform outputs to angle and prepare prediction
+        norm = torch.sqrt(x[:, 0] ** 2 + x[:, 1] ** 2 + x[:, 2] ** 2)
+        kappa = torch.abs(x[:, 3]) + eps_like(x)
+        vec_x = x[:, 0] / norm
+        vec_y = x[:, 1] / norm
+        vec_z = x[:, 2] / norm
+        return torch.stack((vec_x, vec_y, vec_z, kappa), dim=1)
+
+
 class PassOutput1(Task):
     """Passes 1 output without interference."""
 
