### 2022-10-12 8b65b42
Add ModelConfig class
