+++ src/graphnet/models/config.py
@@ -1,0 +1,197 @@
+"""Config classes for the `graphnet.data` module."""
+from functools import wraps
+import inspect
+import types
+from typing import Any, Callable, Dict, Optional, OrderedDict
+
+from pydantic import BaseModel
+import ruamel.yaml as yaml
+
+import graphnet
+from graphnet.models import Model
+from graphnet.utilities.logging import LoggerMixin
+
+
+CONFIG_FILES_SUFFIXES = (".yml", ".yaml")
+
+
+class ModelConfig(BaseModel, LoggerMixin):
+    """Config for data conversion and ingestion."""
+
+    class_name: str
+    arguments: Dict[str, Any]
+
+    def __init__(self, **data: Any) -> None:
+        """Constructor."""
+        # Parse any nested `ModelConfig` arguments
+        for arg in data["arguments"]:
+            value = data["arguments"][arg]
+            if (
+                isinstance(value, dict)
+                and len(value) == 1
+                and self.__class__.__name__ in value
+            ):
+                nested_config_dict = value[self.__class__.__name__]
+                nested_config = self.__class__(**nested_config_dict)
+                data["arguments"][arg] = nested_config
+
+        super().__init__(**data)
+
+    @classmethod
+    def load(cls, path: str) -> "ModelConfig":
+        """Load ModelConfig from `path`."""
+        assert path.endswith(
+            CONFIG_FILES_SUFFIXES
+        ), "Please specify YAML config file."
+        with open(path, "r") as f:
+            config_dict = yaml.load(f)
+
+        return cls(**config_dict)
+
+    def dump(self, path: Optional[str] = None) -> Optional[str]:
+        """Save ModelConfig to `path` as YAML file, or return as string."""
+        config_dict = self._as_dict()[self.__class__.__name__]
+
+        if path:
+            if not path.endswith(CONFIG_FILES_SUFFIXES):
+                path += CONFIG_FILES_SUFFIXES[0]
+            with open(path, "w") as f:
+                yaml.dump(config_dict, f)
+        else:
+            return yaml.dump(config_dict)
+
+    def construct_model(self, trust: bool = False) -> "Model":
+        """Construct model based on current config.
+
+        Arguments:
+            trust (bool): Whether to trust the ModelConfig file enough to
+                `eval(...)` any lambda function expressions contained.
+
+        Raises:
+            ValueError: If the ModelConfig contains lambda functions but
+                `trust = False`.
+        """
+        # Get a lookup for all classes in `graphnet`
+        namespace_classes = get_namespace_classes(graphnet)
+
+        # Parse potential ModelConfig arguments
+        arguments = dict(**self.arguments)
+        for arg, value in arguments.items():
+            if isinstance(value, ModelConfig):
+                arguments[arg] = value.construct_model()
+            elif isinstance(value, str) and value.startswith("!lambda"):
+                if trust:
+                    arguments[arg] = eval(value[1:])
+                else:
+                    self.logger.error(
+                        "Constructing model containing a lambda function "
+                        f"({value}) with `trust=False` if you trust the lambda "
+                        "functions in this ModelConfig, set `trust=True` and "
+                        "reconstruct the model again."
+                    )
+                    raise ValueError
+
+        # Construct model based on arguments
+        return namespace_classes[self.class_name](**arguments)
+
+    def _as_dict(self) -> Dict[str, Dict[str, Any]]:
+        """Represent ModelConfig as a dict.
+
+        This builds on `BaseModel.dict()` but wraps the output in a single-key
+        dictionary to make it unambiguous to identify model arguments that are
+        themselves models.
+        """
+        config_dict = self.dict()
+        for arg, obj in self.arguments.items():
[…]
