### 2022-10-12 d845a75
Introduce base Model class; rename Model -> Standard Model; and track config of all Model instances
