+++ src/graphnet/models/gnn/gnn.py
@@ -4,11 +4,13 @@ from torch import Tensor
 from torch_geometric.data import Data
 
 from graphnet.models import Model
+from graphnet.models.config import save_config
 
 
 class GNN(Model):
     """Base class for all core GNN models in graphnet."""
 
+    @save_config
     def __init__(self, nb_inputs, nb_outputs):
         # Base class constructor
         super().__init__()
