+++ src/graphnet/training/loss_functions.py
@@ -4,6 +4,8 @@ All loss functions inherit from `LossFunction` which (...)
 """
 
 from abc import abstractmethod
+from typing import Optional
+
 
 try:
     from typing import final
@@ -13,17 +15,19 @@ except ImportError:  # Python version < 3.8
         return f
 
 
-from typing import Optional
 import numpy as np
 import scipy.special
 import torch
 from torch import Tensor
-from torch.nn.modules.loss import _WeightedLoss
+
+from graphnet.models.config import save_config
+from graphnet.models.model import Model
 
 
-class LossFunction(_WeightedLoss):
+class LossFunction(Model):
     """Base class for loss functions in graphnet."""
 
+    @save_config
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
