### 2022-10-12 a7ca1e7
Introduce base Model class; rename Model -> Standard Model; and track config of all Model instances
