### 2022-10-12 75cf2db
Better handle nested models and non-graphnet classes
