+++ src/graphnet/models/config.py
@@ -1,8 +1,9 @@
 """Config classes for the `graphnet.data` module."""
+from ast import arguments
 from functools import wraps
 import inspect
 import types
-from typing import Any, Callable, Dict, Optional, OrderedDict
+from typing import Any, Callable, Dict, Union, Optional, OrderedDict
 
 from pydantic import BaseModel
 import ruamel.yaml as yaml
@@ -15,6 +16,23 @@ from graphnet.utilities.logging import LoggerMixin
 CONFIG_FILES_SUFFIXES = (".yml", ".yaml")
 
 
+def traverse_and_apply(
+    obj: Any, fn: Callable, fn_kwargs: Optional[Dict[str, Any]] = None
+):
+    """Apply `fn` to all elements in `obj`, resulting in same structure."""
+    if isinstance(obj, (list, tuple)):
+        return [traverse_and_apply(elem, fn, fn_kwargs) for elem in obj]
+    elif isinstance(obj, dict):
+        return {
+            key: traverse_and_apply(val, fn, fn_kwargs)
+            for key, val in obj.items()
+        }
+    else:
+        if fn_kwargs is None:
+            fn_kwargs = {}
+        return fn(obj, **fn_kwargs)
+
+
 class ModelConfig(BaseModel, LoggerMixin):
     """Config for data conversion and ingestion."""
 
@@ -24,17 +42,35 @@ class ModelConfig(BaseModel, LoggerMixin):
     def __init__(self, **data: Any) -> None:
         """Constructor."""
         # Parse any nested `ModelConfig` arguments
+        def _is_model_config_entry(entry: Dict[str, Any]) -> bool:
+            return (
+                isinstance(entry, dict)
+                and len(entry) == 1
+                and self.__class__.__name__ in entry
+            )
+
+        def _parse_model_config_entry(
+            entry: Dict[str, Any]
+        ) -> Union["ModelConfig", Any]:
+            """Parse dictionary entry to `ModelConfig`."""
+            assert _is_model_config_entry(entry)
+            config_dict = entry[self.__class__.__name__]
+            config = self.__class__(**config_dict)
+            return config
+
         for arg in data["arguments"]:
             value = data["arguments"][arg]
-            if (
-                isinstance(value, dict)
-                and len(value) == 1
-                and self.__class__.__name__ in value
-            ):
-                nested_config_dict = value[self.__class__.__name__]
-                nested_config = self.__class__(**nested_config_dict)
-                data["arguments"][arg] = nested_config
+            if _is_model_config_entry(value):
+                data["arguments"][arg] = _parse_model_config_entry(value)
 
+            elif isinstance(value, (tuple, list)):
+                for ix, elem in enumerate(value):
+                    if _is_model_config_entry(elem):
+                        data["arguments"][arg][ix] = _parse_model_config_entry(
+                            elem
+                        )
+
+        # Base class constructor
         super().__init__(**data)
 
     @classmethod
@@ -44,7 +80,7 @@ class ModelConfig(BaseModel, LoggerMixin):
             CONFIG_FILES_SUFFIXES
         ), "Please specify YAML config file."
         with open(path, "r") as f:
-            config_dict = yaml.load(f)
+            config_dict = yaml.safe_load(f)
 
         return cls(**config_dict)
 
@@ -76,23 +112,65 @@ class ModelConfig(BaseModel, LoggerMixin):
 
         # Parse potential ModelConfig arguments
         arguments = dict(**self.arguments)
-        for arg, value in arguments.items():
-            if isinstance(value, ModelConfig):
-                arguments[arg] = value.construct_model()
-            elif isinstance(value, str) and value.startswith("!lambda"):
-                if trust:
-                    arguments[arg] = eval(value[1:])
-                else:
[…]
