### 2022-10-12 1b5493f
Allow for loading additional modules when parsing lambda functions
