### 2022-10-12 693fc2d
Allow for loading additional modules when parsing lambda functions
