### 2022-10-12 bcd15fa
Rename __call__ -> forward
