### 2022-10-13 bad159f
Call base class constructor in Coarsening
