### 2022-10-14 d57a978
Make LoggerMixin expose .{critical,error,warning,info,debug}() methods directly, rather than accessor to Logger instance
