### 2022-10-14 c8cd30d
Make LoggerMixin expose .{critical,error,warning,info,debug}() methods directly, rather than accessor to Logger instance
