### 2022-10-14 55a310c
Add mypy and docformatter as requirements
