+++ setup.py
@@ -28,7 +28,9 @@ EXTRAS_REQUIRE = {
         "black",
         "colorlog",
         "coverage",
+        "docformatter",
         "MarkupSafe<=2.1",
+        "mypy",
         "pre-commit",
         "pydocstyle",
         "pylint",
