+++ .pre-commit-config.yaml
@@ -9,3 +9,11 @@ repos:
     rev: 4.0.1
     hooks:
     - id: flake8
+  - repo: https://github.com/PyCQA/docformatter
+    rev: v1.5.0
+    hooks:
+    - id: docformatter
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v0.982
+    hooks:
+    - id: mypy
