### 2022-10-14 147c0f7
Update contributing guide with mypy and docformatter
