### 2022-10-14 6212c9e
Add pydocstyle pre-commit hook and update contributing guide accordingly
