### 2022-10-14 03dadc6
Add pydocstyle pre-commit hook and update contributing guide accordingly
