+++ .pre-commit-config.yaml
@@ -9,10 +9,15 @@ repos:
     rev: 4.0.1
     hooks:
     - id: flake8
-  - repo: https://github.com/PyCQA/docformatter
+  - repo: https://github.com/pycqa/docformatter
     rev: v1.5.0
     hooks:
     - id: docformatter
+  - repo: https://github.com/pycqa/pydocstyle
+    rev: 6.1.1
+    hooks:
+    - id: pydocstyle
+      args: [--ignore=D107,D204]
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.982
     hooks:
