### 2022-10-14 b42cc6d
Run docformatter on all files
