### 2022-10-14 7f4ae15
Run docformatter on all files
