+++ .pre-commit-config.yaml
@@ -21,3 +21,4 @@ repos:
     rev: v0.982
     hooks:
     - id: mypy
+      args: [--follow-imports=silent]
