### 2022-10-14 b63b0fc
Remove wildcard imports and add docstring
