### 2022-10-14 ec02540
Update paper.md
Fixing typo
