+++ paper/paper.md
@@ -88,7 +88,7 @@ Finally, the fast (order millisecond) reconstruction allows for a whole new type
 
 `GraphNeT` comprises a number of modules providing the necessary tools to build workflow from ingesting raw training data in domain-specific formats to deploying trained models in domain-specific reconstruction chains, as illustrated in \autoref{fig:flowchart}.
 
-![High-level overview of a typical workflow using `GraphNeT`: `graphnet.data` enables converting domain-specific data to industry-standard, intermediate file formats and reading this data; `graphnet.data` allows for configuring and building complex GNN models using simple, physics-oriented components; `graphnet.training` manages model training and experiment logging; and finally, `graphnet.deployment` allows for using trained models for inference in domain-specific reconstruction chains.\label{fig:flowchart}](flowchart.pdf)
+![High-level overview of a typical workflow using `GraphNeT`: `graphnet.data` enables converting domain-specific data to industry-standard, intermediate file formats and reading this data; `graphnet.models` allows for configuring and building complex GNN models using simple, physics-oriented components; `graphnet.training` manages model training and experiment logging; and finally, `graphnet.deployment` allows for using trained models for inference in domain-specific reconstruction chains.\label{fig:flowchart}](flowchart.pdf)
 
 `graphnet.models` provides modular components subclassing `torch.nn.Module`, meaning that users only need to import a few, existing, purpose-built components and chain them together to form a complete GNN. ML developers can contribute to `GraphNeT` by extending this suite of model components — through new layer types, physics tasks, graph connectivities, etc. — and experiment with optimising these for different reconstruction tasks using experiment tracking.
 
