+++ .pre-commit-config.yaml
@@ -13,6 +13,7 @@ repos:
     rev: v1.5.0
     hooks:
     - id: docformatter
+      args: [--config=setup.cfg]
   - repo: https://github.com/pycqa/pydocstyle
     rev: 6.1.1
     hooks:
