### 2022-10-14 2a69290
Add docstring
