+++ src/graphnet/data/__init__.py
@@ -1,5 +1,4 @@
 """Modules for converting and ingesting data.
 
-`graphnet.data` enables converting domain-specific data to industry-
-standard, intermediate file formats and reading this data.
+`graphnet.data` enables converting domain-specific data to industry-standard, intermediate file formats and reading this data.
 """
