+++ src/graphnet/data/__init__.py
@@ -1,4 +1,5 @@
 """Modules for converting and ingesting data.
 
-`graphnet.data` enables converting domain-specific data to industry-standard, intermediate file formats and reading this data.
+`graphnet.data` enables converting domain-specific data to industry-standard,
+intermediate file formats  and reading this data.
 """
