+++ setup.cfg
@@ -17,3 +17,7 @@ omit =
 [flake8]
 exclude =
     versioneer.py
+
+[docformatter]
+wrap-summaries = 79
+wrap-descriptions = 79
