+++ src/graphnet/training/weight_fitting.py
@@ -8,8 +8,8 @@ from graphnet.data.utils import run_sql_code, save_to_sql, create_table
 class UniformWeightFitter:
     """Produces per-event weights making fitted variable distribution uniform.
 
-    Weights are returned by the public method `fit_weights()`, and the
-    weights can be saved as a table in the database.
+    Weights are returned by the public method `fit_weights()`, and the weights
+    can be saved as a table in the database.
     """
 
     def __init__(
