+++ src/graphnet/deployment/i3modules/__init__.py
@@ -1,8 +1,8 @@
 """Collection of I3Modules.
 
 This module contains various I3Modules which allow for running GNN model
-inference as part of an icetray reconstruction chain, for different
-IceCube detector configurations.
+inference as part of an icetray reconstruction chain, for different IceCube
+detector configurations.
 """
 
 from .graphnet_module import *
