### 2022-10-14 22c31da
Running docformatter on all files with updated settings
