+++ src/graphnet/models/coarsening.py
@@ -83,9 +83,8 @@ class Coarsening(ABC, LoggerMixin):
     def _additional_features(self, cluster: LongTensor, data: Batch) -> Tensor:
         """Additional poolings of feature tensor `x` on `data`.
 
-        By default the nominal `pooling_method` is used for features as
-        well. This method can be overwritten for bespoke coarsening
-        operations.
+        By default the nominal `pooling_method` is used for features as well.
+        This method can be overwritten for bespoke coarsening operations.
         """
 
     def _transfer_attributes(
