### 2022-10-14 1b2788b
Update docstring
