### 2022-10-14 9016c41
Update i3extractor.py
