+++ setup.cfg
@@ -8,7 +8,7 @@ parentdir_prefix = graphnet-
 
 [pydocstyle]
 match = (?!(test_|_version)).*\.py
-convention = google
+ignore = D100,D203,D213,D407,D413
 
 [coverage:run]
 omit =
@@ -21,3 +21,6 @@ exclude =
 [docformatter]
 wrap-summaries = 79
 wrap-descriptions = 79
+
+[mypy]
+ignore_missing_imports = True
