### 2022-10-14 12ae1da
Add type hints
