### 2022-10-14 c394576
Add ignored PEP257 rule D406
