+++ setup.cfg
@@ -8,7 +8,7 @@ parentdir_prefix = graphnet-
 
 [pydocstyle]
 match = (?!(test_|_version)).*\.py
-ignore = D203,D213,D407,D413
+ignore = D203,D213,D406,D407,D413
 
 [coverage:run]
 omit =
