+++ src/graphnet/data/extractors/i3extractor.py
@@ -25,10 +25,10 @@ class I3Extractor(ABC, LoggerMixin):
                 of the provenance of different data, and to name tables to
                 which this data is saved.
         """
-        # Member variables
+        # Member variable(s)
         self._i3_file: Optional[str] = None
         self._gcd_file: Optional[str] = None
-        self._gcd_dict: Optional[Dict[str, Any]] = None
+        self._gcd_dict: Dict[int, Any] = {}
         self._calibration: Optional["icetray.I3Frame.Calibration"] = None
         self._name: str = name
 
