### 2022-10-14 e0fbc3a
Updatge i3genericextractor.py
