+++ src/graphnet/data/extractors/i3splinempeextractor.py
@@ -1,12 +1,28 @@
+"""I3Extractor class(es) for extracting SplineMPE reconstruction."""
+
+from typing import TYPE_CHECKING, Dict
+
 from graphnet.data.extractors.i3extractor import I3Extractor
+
+if TYPE_CHECKING:
+    from icecube import icetray  # pyright: reportMissingImports=false
 
 
 class I3SplineMPEICExtractor(I3Extractor):
-    def __init__(self, name="spline_mpe_ic"):
+    """Class for extracting SplineMPE pointing predictions."""
+
+    def __init__(self, name: str = "spline_mpe_ic"):
+        """Construct instance.
+
+        Args:
+            name (str, optional): Name of the `I3Extractor` instance. Defaults
+                to "spline_mpe_ic".
+        """
+        # Base class constructor
         super().__init__(name)
 
-    def __call__(self, frame) -> dict:
-        """Extracts SplineMPE pointing predictions."""
+    def __call__(self, frame: "icetray.I3Frame") -> Dict[str, float]:
+        """Extract SplineMPE pointing predictions."""
         output = {}
         if "SplineMPEIC" in frame:
             output.update(
