+++ src/graphnet/data/extractors/i3extractor.py
@@ -1,4 +1,5 @@
 """Base I3Extractor class(es)."""
+
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
