### 2022-10-14 48ee695
Update src/graphnet/models/task/reconstruction.py
Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
