+++ src/graphnet/models/task/reconstruction.py
@@ -42,16 +42,15 @@ class AzimuthReconstruction(AzimuthReconstructionWithKappa):
 class DirectionReconstructionWithKappa(Task):
     """Reconstructs Direction with kappa from The 3DVonMisesFisher Distribution."""
 
-    # Requires 4 features: untransformed points in (x,y,z)-space + kappa estimation.
-    nb_inputs = 4
+    # Requires three features: untransformed points in (x,y,z)-space.
+    nb_inputs = 3
 
     def _forward(self, x):
         # Transform outputs to angle and prepare prediction
-        norm = torch.sqrt(x[:, 0] ** 2 + x[:, 1] ** 2 + x[:, 2] ** 2)
-        kappa = torch.abs(x[:, 3]) + eps_like(x)
-        vec_x = x[:, 0] / norm
-        vec_y = x[:, 1] / norm
-        vec_z = x[:, 2] / norm
+        kappa = torch.linalg.vector_norm(x, dim=1) + eps_like(x)
+        vec_x = x[:, 0] / kappa
+        vec_y = x[:, 1] / kappa
+        vec_z = x[:, 2] / kappa
         return torch.stack((vec_x, vec_y, vec_z, kappa), dim=1)
 
 
