### 2022-10-25 55c2261
Add relative DOM efficiency to unit test
