### 2022-10-25 #4831 @dfm
@asogaard — I'm taking over as managing editor here - thanks all for the ping! Like @kthyng says unfortunately the appropriate editors are all still at full capacity, but I'll keep you updated and get someone assigned ASAP. Thanks for your patience!
