### 2022-10-26 #4831 @asogaard
Thanks, @kthyng and @dfm! In that case, I would suggest the following

Potential reviewers within neutrino physics (most relevant to scope of the present paper):
* **JostMigenda**
* **apizzuto** (NB: Alex is also part of the IceCube Collaboration, like the authors of this paper. However, IceCube comprises approx. 300 researchers, and Alex has no direct connection to, or conflicts of interests regarding, the present paper.)

Potential reviewers within astro/-particle physics:
* **bradkav**
* **rafaelab**
* **benjaminrose**
