+++ _sources/graphnet.rst.txt
@@ -8,6 +8,7 @@ Subpackages
    :maxdepth: 4
 
    graphnet.data
+   graphnet.deployment
    graphnet.models
    graphnet.training
    graphnet.utilities
