### 2022-10-26 ae373f4
Update i3truthextractor.py
