### 2022-10-26 f88ae78
Update i3truthextractor.py
