+++ src/graphnet/data/extractors/i3truthextractor.py
@@ -1,6 +1,8 @@
+"""I3Extractor class(es) for extracting truth-level information."""
+
 import numpy as np
 import matplotlib.path as mpath
-from typing import Tuple
+from typing import TYPE_CHECKING, Any, Dict, List, Tuple
 
 from graphnet.data.extractors.i3extractor import I3Extractor
 from graphnet.data.extractors.utilities.frames import (
@@ -9,7 +11,7 @@ from graphnet.data.extractors.utilities.frames import (
 )
 from graphnet.utilities.imports import has_icecube_package
 
-if has_icecube_package():
+if has_icecube_package() or TYPE_CHECKING:
     from icecube import (
         dataclasses,
         icetray,
@@ -18,8 +20,22 @@ if has_icecube_package():
 
 
 class I3TruthExtractor(I3Extractor):
-    def __init__(self, name="truth", borders=None):
+    """Class for extracting truth-level information."""
+
+    def __init__(self, name: str = "truth", borders: List[np.ndarray] = None):
+        """Construct instance.
+
+        Args:
+            name (str, optional): Name of the `I3Extractor` instance. Defaults
+                to "truth".
+            borders (List[np.ndarray], optional): Array of boundaries of the
+                detector volume as ((x,y),z)-coordinates, for identifying,
+                e.g., particles starting and stopping within the detector.
+                Defaults to None.
+        """
+        # Base class constructor
         super().__init__(name)
+
         if borders is None:
             border_xy = np.array(
                 [
@@ -58,8 +74,10 @@ class I3TruthExtractor(I3Extractor):
         else:
             self._borders = borders
 
-    def __call__(self, frame, padding_value=-1) -> dict:
-        """Extracts truth features."""
+    def __call__(
+        self, frame: "icetray.I3Frame", padding_value: Any = -1
+    ) -> Dict[str, Any]:
+        """Extract truth-level information."""
         is_mc = frame_is_montecarlo(frame)
         is_noise = frame_is_noise(frame)
         sim_type = self._find_data_type(is_mc, self._i3_file)
@@ -155,8 +173,7 @@ class I3TruthExtractor(I3Extractor):
                     inelasticity,
                 ) = self._get_primary_track_energy_and_inelasticity(frame)
             except RuntimeError:  # track energy fails on northeren tracks with ""Hadrons" has no mass implemented. Cannot get total energy."
-                energy_track = (padding_value,)
-                inelasticity = (padding_value,)
+                energy_track, inelasticity = (padding_value, padding_value)
 
             output.update(
                 {
@@ -196,7 +213,9 @@ class I3TruthExtractor(I3Extractor):
 
         return output
 
-    def _extract_dbang_decay_length(self, frame, padding_value):
+    def _extract_dbang_decay_length(
+        self, frame: "icetray.I3Frame", padding_value: float = -1
+    ) -> float:
         mctree = frame["I3MCTree"]
         try:
             p_true = mctree.primaries[0]
@@ -237,23 +256,36 @@ class I3TruthExtractor(I3Extractor):
             return padding_value
 
     def _muon_stopped(
-        self, truth, borders, horizontal_pad=100.0, vertical_pad=100.0
-    ):
-        """
-        Calculates where a simulated muon stops and if this is inside the detectors fiducial volume.
-        IMPORTANT: The final position of the muon is saved in truth extractor/databases as position_x,position_y and position_z.
-                This is analogoues to the neutrinos whose interaction vertex is saved under the same name.
+        self,
+        truth: Dict[str, Any],
+        borders: List[np.ndarray],
+        shrink_horizontally: float = 100.0,
+        shrink_vertically: float = 100.0,
+    ) -> Dict[str, Any]:
[…]
