+++ src/graphnet/data/extractors/i3tumextractor.py
@@ -1,12 +1,28 @@
+"""I3Extractor class(es) for extracting TUM DNN reconstruction."""
+
+from typing import TYPE_CHECKING, Dict
+
 from graphnet.data.extractors.i3extractor import I3Extractor
+
+if TYPE_CHECKING:
+    from icecube import icetray  # pyright: reportMissingImports=false
 
 
 class I3TUMExtractor(I3Extractor):
-    def __init__(self, name="tum_dnn"):
+    """Class for extracting TUM DNN predictions."""
+
+    def __init__(self, name: str = "tum_dnn"):
+        """Construct instance.
+
+        Args:
+            name (str, optional): Name of the `I3Extractor` instance. Defaults
+                to "tum_dnn".
+        """
+        # Base class constructor
         super().__init__(name)
 
-    def __call__(self, frame) -> dict:
-        """Extracts TUM DNN Recos and associated variables."""
+    def __call__(self, frame: "icetray.I3Frame") -> Dict[str, float]:
+        """Extract TUM DNN recoconstruction and associated variables."""
         output = {}
         if "TUM_dnn_energy_hive" in frame:
             output.update(
