+++ src/graphnet/data/parquet/parquet_dataconverter.py
@@ -1,3 +1,5 @@
+"""DataConverter for the Parquet backend."""
+
 from collections import OrderedDict
 import os
 from typing import List, Optional
@@ -8,9 +10,10 @@ from graphnet.data.dataconverter import DataConverter
 
 
 class ParquetDataConverter(DataConverter):
+    """Class for converting I3-files to Parquet format."""
 
     # Class variables
-    file_suffix = "parquet"
+    file_suffix: str = "parquet"
 
     # Abstract method implementation(s)
     def save_data(self, data: List[OrderedDict], output_file: str):
@@ -34,4 +37,17 @@ class ParquetDataConverter(DataConverter):
     def merge_files(
         self, output_file: str, input_files: Optional[List[str]] = None
     ):
+        """Parquet-specific method for merging output files.
+
+        Args:
+            output_file (str): Name of the output file containing the merged
+                results.
+            input_files (List[str]): Intermediate files to be merged, according
+                to the specific implementation. Default to None, meaning that
+                all files output by the current instance are merged.
+
+        Raises:
+            NotImplementedError: If the method has not been implemented for the
+                Parquet backend.
+        """
         raise NotImplementedError()
