### 2022-10-26 a100069
Add output type hint
