+++ src/graphnet/data/parquet/parquet_dataset.py
@@ -34,7 +34,7 @@ class ParquetDataset(Dataset):
         # Set custom member variable(s)
         self._parquet_hook = ak.from_parquet(self._path, lazy=False)
 
-    def _get_all_indices(self):
+    def _get_all_indices(self) -> List[int]:
         return ak.to_numpy(
             self._parquet_hook[self._truth_table][self._index_column]
         ).tolist()
