### 2022-10-26 650dd98
Add docstring
