+++ src/graphnet/data/sqlite/__init__.py
@@ -1,3 +1,5 @@
+"""SQLite-specific implementation of data classes."""
+
 from graphnet.utilities.imports import has_torch_package
 
 from .sqlite_dataconverter import SQLiteDataConverter
