### 2022-10-26 aed9ac6
Add colorlog as required dependency
