### 2022-10-26 9c202e0
Update docstrings
