+++ src/graphnet/data/dataset.py
@@ -33,6 +33,49 @@ class Dataset(ABC, torch.utils.data.Dataset, LoggerMixin):
         loss_weight_column: str = None,
         loss_weight_default_value: Optional[float] = None,
     ):
+        """Construct Dataset.
+
+        Args:
+            path: Path to the file(s) from which this `Dataset` should read.
+            pulsemaps: Name(s) of the pulse map series that should be used to
+                construct the nodes on the individual graph objects, and their
+                features. Multiple pulse series maps can be used, e.g., when
+                different DOM types are stored in different maps.
+            features: List of columns in the input files that should be used as
+                node features on the graph objects.
+            truth: List of event-level columns in the input files that should
+                be used added as attributes on the  graph objects.
+            node_truth: List of node-level columns in the input files that
+                should be used added as attributes on the graph objects.
+                Defaults to None.
+            index_column: Name of the column in the input files that contains
+                unique indicies to identify and map events across tables.
+                Defaults to "event_no".
+            truth_table: Name of the table containing event-level truth
+                information. Defaults to "truth".
+            node_truth_table: Name of the table containing node-level truth
+                information. Defaults to None.
+            string_selection: Subset of strings for which data should be read
+                and used to construct graph objects. Defaults to None, meaning
+                all strings for which data exists are used.
+            selection: List of indicies (in `index_column`) of the events in
+                the input files that should be read. Defaults to None, meaning
+                that all events in the input files are read.
+            dtype: Type of the feature tensor on the graph objects returned.
+                Defaults to torch.float32.
+            loss_weight_table: Name of the table containing per-event loss
+                weights. Defaults to None.
+            loss_weight_column: Name of the column in `loss_weight_table`
+                containing per-event loss weights. This is also the name of the
+                corresponding attribute assigned to the graph object. Defaults
+                to None.
+            loss_weight_default_value : Default per-event loss weight.
+                NOTE: This default value is only applied when
+                `loss_weight_table` and `loss_weight_column` are specified, and
+                in this case to events with no value in the corresponding
+                table/column. That is, if no per-event loss weight table/column
+                is provided, this value is ignored. Defaults to None.
+        """
         # Check(s)
         if isinstance(pulsemaps, str):
             pulsemaps = [pulsemaps]
@@ -102,7 +145,7 @@ class Dataset(ABC, torch.utils.data.Dataset, LoggerMixin):
             self._indices = selection
 
         # Purely internal member variables
-        self._missing_variables = {}
+        self._missing_variables: Dict[str, List[str]] = {}
         self._remove_missing_columns()
 
         # Implementation-specific post-init code.
@@ -127,22 +170,22 @@ class Dataset(ABC, torch.utils.data.Dataset, LoggerMixin):
         columns: Union[List[str], str],
         index: int,
         selection: Optional[str] = None,
-    ) -> List[Tuple[Any]]:
+    ) -> List[Tuple[Any, ...]]:
         """Query a table at a specific index, optionally with some selection.
 
         Args:
-            table (str): Table to be queried.
-            columns (List[str]): Columns to read out.
-            index (int): Sequentially numbered index (i.e. in [0,len(self))) of
-                the event to query. This _may_ differ from the indexation used
-                in `self._indices`.
-            selection (Optional[str], optional): Selection to be imposed before
-                reading out data. Defaults to None.
+            table: Table to be queried.
+            columns: Columns to read out.
[…]
