### 2022-10-26 7a99fd6
Update title of API docs and auto-detect type hints
