+++ src/graphnet/data/utilities/random.py
@@ -1,5 +1,4 @@
-"""Random generator-related utility functions relevant to the graphnet.data
-package."""
+"""RNG-related utility functions relevant to the graphnet.data package."""
 
 from typing import List, Tuple
 import pandas as pd
@@ -7,18 +6,18 @@ import pandas as pd
 
 def pairwise_shuffle(
     i3_list: List[str], gcd_list: List[str]
-) -> Tuple[List[str]]:
-    """Shuffles the I3 file list and the correponding gcd file list.
+) -> Tuple[List[str], List[str]]:
+    """Shuffle the I3 file list and the correponding gcd file list.
 
     This is handy because it ensures a more even extraction load for each worker.
 
     Args:
-        files_list (list): List of I3 file paths.
-        gcd_list (list): List of corresponding gcd file paths.
+        files_list: List of I3 file paths.
+        gcd_list: List of corresponding gcd file paths.
 
     Returns:
-        i3_shuffled (list): List of shuffled I3 file paths.
-        gcd_shuffled (list): List of corresponding gcd file paths.
+        i3_shuffled: List of shuffled I3 file paths.
+        gcd_shuffled: List of corresponding gcd file paths.
     """
     df = pd.DataFrame({"i3": i3_list, "gcd": gcd_list})
     df_shuffled = df.sample(frac=1, replace=False)
