+++ src/graphnet/data/dataconverter.py
@@ -1,21 +1,37 @@
+# type: ignore[name-defined]
+"""Base `DataConverter` class(es) used in GraphNeT."""
+
 from abc import ABC, abstractmethod
 from collections import OrderedDict
 from dataclasses import dataclass
 from functools import wraps
 import itertools
 from multiprocessing import Manager, Pool, Value
+import multiprocessing.pool
+from multiprocessing.sharedctypes import Synchronized
 import os
 import re
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    List,
+    Optional,
+    Tuple,
+    TypeVar,
+    Union,
+    cast,
+)
+
 import numpy as np
 import pandas as pd
-from typing import Callable, List, Optional, Tuple, Union
 from tqdm import tqdm
 
 try:
     from typing import final
 except ImportError:  # Python version < 3.8
 
-    def final(f):  # Identity decorator
+    def final(f):  # Identity decorator  # noqa: D103
         return f
 
 
@@ -44,21 +60,23 @@ SAVE_STRATEGIES = [
 
 # Utility classes
 @dataclass
-class FileSet:
+class FileSet:  # noqa: D101
     i3_file: str
     gcd_file: str
 
 
 # Utility method(s)
-def init_global_index(index: Value, output_files: List[str]):
+def init_global_index(index: Synchronized, output_files: List[str]):
     """Make `global_index` available to pool workers."""
     global global_index, global_output_files
-    global_index = index
-    global_output_files = output_files
+    global_index, global_output_files = (index, output_files)  # type: ignore[name-defined]
 
 
-def cache_output_files(process_method):
-    """Decorator to cache output file names."""
+F = TypeVar("F", bound=Callable[..., Any])
+
+
+def cache_output_files(process_method: F) -> F:
+    """Decorate `process_method` to cache output file names."""
 
     @wraps(process_method)
     def wrapper(self, *args):
@@ -73,12 +91,11 @@ def cache_output_files(process_method):
         output_files.append(output_file)
         return output_file
 
-    return wrapper
+    return cast(F, wrapper)
 
 
 class DataConverter(ABC, LoggerMixin):
-    """Abstract base class for specialised (SQLite, parquet, etc.)
-    converters."""
+    """Base class for converting I3-files to intermediate file format."""
 
     @property
     @abstractmethod
@@ -98,7 +115,7 @@ class DataConverter(ABC, LoggerMixin):
         index_column: str = "event_no",
         icetray_verbose: int = 0,
     ):
-        """Converts I3 files to an intermediate format.
+        """Construct DataConverter.
 
         When using `input_file_batch_pattern`, regular expressions are used to
         group files according to their names. All files that match a certain
@@ -171,7 +188,7 @@ class DataConverter(ABC, LoggerMixin):
         # Placeholders for keeping track of sequential event indices and output files
         self._index_column = index_column
         self._index = 0
-        self._output_files = []
+        self._output_files: List[str] = []
 
         # Set verbosity
         if icetray_verbose == 0:
@@ -179,12 +196,11 @@ class DataConverter(ABC, LoggerMixin):
 
     @final
     def __call__(self, directories: Union[str, List[str]]):
-        """Main call to convert I3 files in `directories.
+        """Convert I3-files in `directories.
 
         Args:
-            directories (Union[str, List[str]]): One or more directories, the I3
-                files within which should be converted to an intermediate file
-                format.
+            directories: One or more directories, the I3 files within which
+                should be converted to an intermediate file format.
         """
         # Find all I3 and GCD files in the specified directories.
         i3_files, gcd_files = find_i3_files(directories, self._gcd_rescue)
@@ -213,9 +229,8 @@ class DataConverter(ABC, LoggerMixin):
         intermediate file format.
 
         Args:
[…]
