### 2022-10-26 ba36592
Remove redundant type information
