### 2022-10-26 bc9ecc7
Remove redundant type information
