### 2022-10-26 a76a91a
Update docstring
