### 2022-10-26 be232e2
Add loss weights
