### 2022-10-26 b88d3b0
Update examples/bjørn_low_weight_fitter_example.py
Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
