### 2022-10-26 a3e12c4
Update src/graphnet/training/weight_fitting.py
Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
