### 2022-10-26 1bb01c8
Update src/graphnet/training/weight_fitting.py
Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
