### 2022-10-27 ade7d47
Multiclass classification implementation
