### 2022-10-27 e1ab4a5
assigned tensor to device
