+++ src/graphnet/training/loss_functions.py
@@ -119,8 +119,10 @@ class MultiClassificationCrossEntropyLoss(LossFunction):
     """
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
-        pid_transform = {1:0,12:2,13:1,14:2,16:2}
-        target_new = one_hot(torch.tensor([pid_transform[np.abs(int(value))] for value in target]), 3)
+        # transformer for multiple pid types to total of three classes
+        device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        pid_transform = {1:0, 12:2, 13:1, 14:2, 16:2}
+        target_new = one_hot(torch.tensor([pid_transform[np.abs(int(value))] for value in target]), 3).to(device)
         return cross_entropy(
             prediction.float(), target_new.float(), reduction="none"
         )
