### 2022-10-28 ace2460
Update docstrings
