+++ src/graphnet/data/extractors/utilities/collections.py
@@ -1,4 +1,4 @@
-"""Utility methods to manipulate python-type collections of data."""
+"""Utility methods for manipulating python-type collections of data."""
 
 # Utility function(s)
 import json
@@ -12,7 +12,7 @@ def flatten_nested_dictionary(
     parent_key: str = "",
     separator: str = "__",
 ) -> Union[Dict, Any]:
-
+    """Flatten a nested dictionary to a dictionary with non-dict values."""
     # Dict-like
     if isinstance(obj, dict):
         items: List[Tuple[str, Any]] = []
@@ -57,7 +57,7 @@ def serialise(obj: Union[Dict, Any]) -> Union[Dict, Any]:
 def transpose_list_of_dicts(
     array: List[Dict[str, Any]]
 ) -> Dict[str, List[Any]]:
-
+    """Transpose a list of dicts to a dict of lists."""
     if len(array) == 0:
         return {}
 
