### 2022-10-28 e0e6a02
Remove assert-statement that is no longer valid after changing definition and type of truth
