+++ src/graphnet/data/extractors/utilities/collections.py
@@ -12,7 +12,18 @@ def flatten_nested_dictionary(
     parent_key: str = "",
     separator: str = "__",
 ) -> Union[Dict, Any]:
-    """Flatten a nested dictionary to a dictionary with non-dict values."""
+    """Flatten a nested dictionary to a dictionary with non-dict values.
+
+    Example:
+        d = {"a": {"b": 1}, "c": 2}
+        flatten_nested_dictionary(d)
+        >>> {"a__b": 1, "c": 2}
+
+    Args:
+        obj: The object that should be flattened, if applicable.
+        parent_key: The combined name of the parent key(s) containing `obj`.
+        separator: The string used to concatenate nester parent keys.
+    """
     # Dict-like
     if isinstance(obj, dict):
         items: List[Tuple[str, Any]] = []
