+++ src/graphnet/data/constants.py
@@ -1,7 +1,9 @@
-"""Global constants that are used across modules."""
+"""Global constants that are used with `graphnet.data`."""
 
 
 class FEATURES:
+    """Namespace for standard names working with `I3FeatureExtractor`."""
+
     ICECUBE86 = [
         "dom_x",
         "dom_y",
@@ -24,6 +26,8 @@ class FEATURES:
 
 
 class TRUTH:
+    """Namespace for standard names working with `I3TruthExtractor`."""
+
     ICECUBE86 = [
         "energy",
         "energy_track",
