+++ src/graphnet/utilities/logging.py
@@ -1,4 +1,4 @@
-"""Consistent and configurable logging across the project."""
+"""Consistent and configurable logging across `graphnet`."""
 
 from collections import Counter
 from functools import lru_cache
@@ -30,7 +30,6 @@ def set_logging_level(level: int = logging.INFO):
 
 def get_formatters() -> Tuple[logging.Formatter, colorlog.ColoredFormatter]:
     """Get coloured and non-coloured logging formatters."""
-
     # Common configuration
     colorlog_format = (
         "\033[1;34m%(name)s\033[0m: "
@@ -64,15 +63,18 @@ class RepeatFilter(object):
     """Filter out repeat messages."""
 
     def __init__(self):
+        """Construct `RepeatFilter`."""
         self._messages = Counter()
         self.nb_repeats_allowed = 20
 
     def filter(self, record):
+        """Filter messages printed more than `nb_repeats_allowed` times."""
         self._messages[record.msg] += 1
         count = self._messages[record.msg]
         if count == self.nb_repeats_allowed:
             get_logger().debug(
-                f"Will not print the below message again ({self.nb_repeats_allowed} repeats reached)."
+                "Will not print the below message again "
+                f"({self.nb_repeats_allowed} repeats reached)."
             )
 
         return count <= self.nb_repeats_allowed
@@ -143,6 +145,8 @@ def get_logger(
 
 
 class LoggerMixin(object):
+    """Class for enabling logging directly from inheriting classes."""
+
     def _get_logger(self):
         """Construct Logger instance if not already done."""
         if not hasattr(self, "_logger"):
