+++ src/graphnet/utilities/filesys.py
@@ -1,14 +1,12 @@
 """File system-related utility functions relevant to the graphnet package."""
 
-from glob import glob
-import os
 from pathlib import Path
 import re
-from typing import List, Optional, Union
+from typing import List, Optional, Union, cast
 
 
 def is_gcd_file(filename: str) -> bool:
-    """Checks whether `filename` is a GCD file."""
+    """Check whether `filename` is a GCD file."""
     if (
         re.search("(gcd|geo)", filename.lower())
         or "GeoCalibDetector" in filename
@@ -18,7 +16,7 @@ def is_gcd_file(filename: str) -> bool:
 
 
 def is_i3_file(filename: str) -> bool:
-    """Checks whether `filename` is an I3 file."""
+    """Check whether `filename` is an I3 file."""
     if is_gcd_file(filename):
         return False
     elif has_extension(filename, ["bz2", "zst", "gz"]):
@@ -27,7 +25,7 @@ def is_i3_file(filename: str) -> bool:
 
 
 def has_extension(filename: str, extensions: List[str]) -> bool:
-    """Checks whether `filename` has one of the desired extensions."""
+    """Check whether `filename` has one of the desired extensions."""
     # @TODO: Remove method, as it is not used?
     return re.search("(" + "|".join(extensions) + ")$", filename) is not None
 
@@ -35,7 +33,7 @@ def has_extension(filename: str, extensions: List[str]) -> bool:
 def find_i3_files(
     directories: Union[str, List[str]], gcd_rescue: Optional[str] = None
 ):
-    """Finds I3 files and corresponding GCD files in `directories`.
+    """Find I3 files and corresponding GCD files in `directories`.
 
     Finds I3 files in dir and matches each file with a corresponding GCD file if
     present in the directory, matches with gcd_rescue if gcd is not present in
@@ -74,7 +72,9 @@ def find_i3_files(
                 str(path) for path in paths if path.parent == folder
             ]
             folder_i3_files = list(filter(is_i3_file, folder_files))
-            folder_gcd_files = list(filter(is_gcd_file, folder_files))
+            folder_gcd_files = cast(
+                List[Optional[str]], list(filter(is_gcd_file, folder_files))
+            )
 
             # Make sure that no more than one GCD file is found; and use rescue file of none is found.
             assert len(folder_gcd_files) <= 1
