### 2022-10-28 d65887b
Update loss_functions.py
