+++ src/graphnet/training/loss_functions.py
@@ -1,6 +1,8 @@
+# type: ignore[name-defined]  # Due to use of `init_global_index`.
 """Collection of loss functions.
 
-All loss functions inherit from `LossFunction` which (...)
+All loss functions inherit from `LossFunction` which ensures a common syntax,
+handles per-event weights, etc.
 """
 
 from abc import abstractmethod
@@ -11,7 +13,7 @@ try:
     from typing import final
 except ImportError:  # Python version < 3.8
 
-    def final(f):  # Identity decorator
+    def final(f):  # Identity decorator  # noqa: D103
         return f
 
 
@@ -25,10 +27,11 @@ from graphnet.models.model import Model
 
 
 class LossFunction(Model):
-    """Base class for loss functions in graphnet."""
+    """Base class for loss functions in `graphnet`."""
 
     @save_config
     def __init__(self, **kwargs):
+        """Construct `LossFunction`, saving model config."""
         super().__init__(**kwargs)
 
     @final
@@ -40,6 +43,7 @@ class LossFunction(Model):
         return_elements: bool = False,
     ) -> Tensor:
         """Forward pass for all loss functions.
+
         Args:
             prediction (Tensor): Tensor containing predictions. Shape [N,P]
             target (Tensor): Tensor containing targets. Shape [N,T]
@@ -62,15 +66,14 @@ class LossFunction(Model):
 
     @abstractmethod
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
-        """Syntax similar to `.forward` for implentation in inheriting
-        classes."""
+        """Syntax like `.forward`, for implentation in inheriting classes."""
 
 
 class MSELoss(LossFunction):
     """Mean squared error loss."""
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
-        """Implementation of loss calculation."""
+        """Implement loss calculation."""
         # Check(s)
         assert prediction.dim() == 2
         assert prediction.size() == target.size()
@@ -83,7 +86,7 @@ class RMSELoss(MSELoss):
     """Root mean squared error loss."""
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
-        """Implementation of loss calculation."""
+        """Implement loss calculation."""
         # Check(s)
         elements = super()._forward(prediction, target)
         elements = torch.sqrt(elements)
@@ -106,19 +109,17 @@ class LogCoshLoss(LossFunction):
         return x + torch.nn.functional.softplus(-2.0 * x) - np.log(2.0)
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
-        """Implementation of loss calculation."""
+        """Implement loss calculation."""
         diff = prediction - target
         elements = self._log_cosh(diff)
         return elements
 
 
 class BinaryCrossEntropyLoss(LossFunction):
-    """Computes binary cross entropy for a vector of predictions (between 0 and
-    1), targets should be 0 and 1 for muon and neutrino respectively where
-    prediction is prob.
+    """Compute binary cross entropy loss.
 
-    the PID is neutrino (12,14,16) loss should be reported elementwise, so set
-    reduction to None
+    Predictions are vector probabilities (i.e., values between 0 and 1), and
+    targets should be 0 and 1.
     """
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
@@ -162,6 +163,7 @@ class LogCMK(torch.autograd.Function):
     def forward(
         ctx, m, kappa
     ):  # pylint: disable=invalid-name,arguments-differ
+        """Forward pass."""
         dtype = kappa.dtype
         ctx.save_for_backward(kappa)
         ctx.m = m
[…]
