### 2022-10-28 43dcc74
Update loss_functions.py
