+++ src/graphnet/data/sqlite/__init__.py
@@ -3,7 +3,7 @@
 from graphnet.utilities.imports import has_torch_package
 
 from .sqlite_dataconverter import SQLiteDataConverter
-from .sqlite_utilities import run_sql_code, save_to_sql
+from .sqlite_utilities import run_sql_code, save_to_sql, create_table
 
 if has_torch_package():
     from .sqlite_dataset import SQLiteDataset
