+++ src/graphnet/data/dataconverter.py
@@ -1,5 +1,4 @@
-# type: ignore[name-defined]  # Due to use of `init_global_index` and override
-# of `final` with different signature.
+# type: ignore[name-defined]  # Due to use of `init_global_index`.
 """Base `DataConverter` class(es) used in GraphNeT."""
 
 from abc import ABC, abstractmethod
@@ -32,7 +31,8 @@ try:
     from typing import final
 except ImportError:  # Python version < 3.8
 
-    def final(f):  # Identity decorator  # noqa: D103
+    # Identity decorator
+    def final(f):  # type: ignore  # noqa: D103
         return f
 
 
