### 2022-10-28 aa1c9ac
Update coarsening.py
