+++ src/graphnet/models/coarsening.py
@@ -1,4 +1,4 @@
-"""Classes for coarsening operations (i.e., clustering, or local pooling."""
+"""Class(es) for coarsening operations (i.e., clustering, or local pooling)."""
 
 from abc import abstractmethod
 from typing import List, Optional, Union
@@ -32,6 +32,7 @@ from torch_geometric.utils import degree
 
 
 def unbatch_edge_index(edge_index: Tensor, batch: Tensor) -> List[Tensor]:
+    # noqa: D401
     r"""Splits the :obj:`edge_index` according to a :obj:`batch` vector.
 
     Args:
@@ -66,6 +67,7 @@ class Coarsening(Model):
         reduce: str = "avg",
         transfer_attributes: bool = True,
     ):
+        """Construct `Coarsening`."""
         assert reduce in self.reduce_options
 
         (
@@ -79,11 +81,10 @@ class Coarsening(Model):
 
     @abstractmethod
     def _perform_clustering(self, data: Union[Data, Batch]) -> LongTensor:
-        """Perform clustering of nodes in `data` by assigning unique cluster
-        indices to each."""
+        """Cluster nodes in `data` by assigning a cluster index to each."""
 
     def _additional_features(self, cluster: LongTensor, data: Batch) -> Tensor:
-        """Additional poolings of feature tensor `x` on `data`.
+        """Perform additional poolings of feature tensor `x` on `data`.
 
         By default the nominal `pooling_method` is used for features as well.
         This method can be overwritten for bespoke coarsening operations.
@@ -101,7 +102,7 @@ class Coarsening(Model):
         batch: Optional[LongTensor] = original_data.batch
         for ix, attr in enumerate(attributes):
             if attr not in pooled_data._store:
-                values = getattr(original_data, attr)
+                values: Tensor = getattr(original_data, attr)
 
                 attr_is_node_level_tensor = False
                 if isinstance(values, Tensor):
@@ -115,7 +116,7 @@ class Coarsening(Model):
                         )
 
                 if attr_is_node_level_tensor:
-                    values: Tensor = self._attribute_reduce_method(
+                    values = self._attribute_reduce_method(
                         cluster,
                         values,
                         batch=torch.zeros_like(values, dtype=torch.int32),
@@ -126,8 +127,7 @@ class Coarsening(Model):
         return pooled_data
 
     def forward(self, data: Union[Data, Batch]) -> Union[Data, Batch]:
-        """Coarsening operation."""
-
+        """Perform coarsening operation."""
         # Get tensor of cluster indices for each node.
         cluster: LongTensor = self._perform_clustering(data)
 
@@ -194,9 +194,10 @@ class Coarsening(Model):
 
 
 class DOMCoarsening(Coarsening):
+    """Coarsen pulses to DOM-level."""
+
     def _perform_clustering(self, data: Union[Data, Batch]) -> LongTensor:
-        """Perform clustering of nodes in `data` by assigning unique cluster
-        indices to each."""
+        """Cluster nodes in `data` by assigning a cluster index to each."""
         # dom_index = group_pulses_to_dom(data)
         dom_index = group_by(
             data, ["dom_x", "dom_y", "dom_z", "rde", "pmt_area"]
@@ -205,8 +206,10 @@ class DOMCoarsening(Coarsening):
 
 
 class CustomDOMCoarsening(DOMCoarsening):
+    """Coarsen pulses to DOM-level with additional attributes."""
+
     def _additional_features(self, cluster: LongTensor, data: Data) -> Tensor:
-        """Additional poolings of feature tensor `x` on `data`."""
[…]
