### 2022-10-28 5a6b930
Remove loop-based coarsening
