+++ src/graphnet/models/coarsening.py
@@ -280,43 +280,3 @@ class DOMAndTimeWindowCoarsening(Coarsening):
         )
 
         return clusters
-
-
-class LoopBasedCoarsening:
-    """Coarsen pulses to DOM-level using for-loop.
-
-    DEPRECATED.
-    """
-
-    def __call__(self, data: Data) -> Data:
-        """Perform coarsening to DOM-level."""
-        unique_doms, inverse_idx, n_pulses_pr_dom = torch.unique(
-            data.x[:, [0, 1, 2, 5, 6]],
-            return_counts=True,
-            return_inverse=True,
-            dim=0,
-        )
-        unique_inverse_indices = torch.unique(inverse_idx)
-        count = 0
-        pulse_statistics = torch.zeros(size=(len(unique_doms), 8))
-
-        # 'dom_x','dom_y','dom_z','dom_time','charge','rde','pmt_area'
-        for unique_inverse_idx in unique_inverse_indices:
-            time = data.x[inverse_idx == unique_inverse_idx, 3]
-            charge = data.x[inverse_idx == unique_inverse_idx, 4]
-            pulse_statistics[count, 0] = torch.min(time)
-            pulse_statistics[count, 1] = torch.mean(time)
-            pulse_statistics[count, 2] = torch.max(time)
-            pulse_statistics[count, 3] = torch.std(time, unbiased=False)
-            pulse_statistics[count, 4] = torch.min(charge)
-            pulse_statistics[count, 5] = torch.mean(charge)
-            pulse_statistics[count, 6] = torch.max(charge)
-            pulse_statistics[count, 7] = torch.std(charge, unbiased=False)
-            count += 1
-
-        data = data.clone()  # @TODO: To avoid modifying in-place?
-        data.x = torch.cat(
-            (unique_doms, n_pulses_pr_dom.unsqueeze(1), pulse_statistics),
-            dim=1,
-        )
-        return data
