### 2022-10-28 ee9750d
Update docstring
