+++ src/graphnet/models/components/layers.py
@@ -1,3 +1,5 @@
+"""Class(es) implementing layers to be used in `graphnet` models."""
+
 from typing import Callable, List, Optional, Sequence, Union
 
 from torch.functional import Tensor
@@ -8,14 +10,28 @@ from torch_geometric.typing import Adj
 
 
 class DynEdgeConv(EdgeConv):
+    """Dynamical edge convolution layer."""
+
     def __init__(
         self,
         nn: Callable,
         aggr: str = "max",
         nb_neighbors: int = 8,
-        features_subset: Optional[Union[Sequence[int], List[int]]] = None,
+        features_subset: Optional[Union[Sequence[int], slice]] = None,
         **kwargs,
     ):
+        """Construct `DynEdgeConv`.
+
+        Args:
+            nn: The MLP/torch.Module to be used within the `EdgeConv`.
+            aggr: Aggregation method to be used with `EdgeConv`.
+            nb_neighbors: Number of neighbours to be clustered after the
+                `EdgeConv` operation.
+            features_subset: Subset of features in `Data.x` that should be used
+                when dynamically performing the new graph clustering after the
+                `EdgeConv` operation. Defaults to all features.
+            **kwargs: Additional features to be passed to `EdgeConv`.
+        """
         # Check(s)
         if features_subset is None:
             features_subset = slice(None)  # Use all features
@@ -31,6 +47,7 @@ class DynEdgeConv(EdgeConv):
     def forward(
         self, x: Tensor, edge_index: Adj, batch: Optional[Tensor] = None
     ) -> Tensor:
+        """Forward pass."""
         # Standard EdgeConv forward pass
         x = super().forward(x, edge_index)
 
