+++ src/graphnet/models/components/pool.py
@@ -1,4 +1,6 @@
-from typing import Any, List, Optional, Union
+"""Functions for performing pooling/clustering/coarsening."""
+
+from typing import Any, Callable, List, Optional, Union
 
 import torch
 from torch import LongTensor, Tensor
@@ -15,8 +17,13 @@ from torch_geometric.nn.pool import (
 )
 
 
-def min_pool(cluster: Any, data: Any, transform: Optional[Any] = None):
-    """Like `max_pool, just negating `data.x`."""
+def min_pool(
+    cluster: LongTensor, data: Data, transform: Optional[Any] = None
+) -> Data:
+    """Perform min-pooling of `Data`.
+
+    Like `max_pool, just negating `data.x`.
+    """
     data.x = -data.x
     data_pooled = max_pool(
         cluster,
@@ -28,8 +35,16 @@ def min_pool(cluster: Any, data: Any, transform: Optional[Any] = None):
     return data_pooled
 
 
-def min_pool_x(cluster: Any, x: Any, batch: Any, size: Optional[int] = None):
-    """Like `max_pool_x, just negating `x`."""
+def min_pool_x(
+    cluster: LongTensor,
+    x: Tensor,
+    batch: LongTensor,
+    size: Optional[int] = None,
+) -> Tensor:
+    """Perform min-pooling of `Tensor`.
+
+    Like `max_pool_x, just negating `x`.
+    """
     ret = max_pool_x(cluster, -x, batch, size)
     if size is None:
         return (-ret[0], ret[1])
@@ -42,8 +57,7 @@ def sum_pool_and_distribute(
     cluster_index: LongTensor,
     batch: Optional[LongTensor] = None,
 ) -> Tensor:
-    """Sum-pool values across the cluster, and distribute the individual
-    nodes."""
+    """Sum-pool values and distribute result to the individual nodes."""
     if batch is None:
         batch = torch.zeros(tensor.size(dim=0)).long()
     tensor_pooled, _ = sum_pool_x(cluster_index, tensor, batch)
@@ -58,12 +72,11 @@ def _group_identical(
     """Group rows in `tensor` that are identical.
 
     Args:
-        tensor (Tensor): Tensor of shape [N, F]
-        batch (Optional[LongTensor], optional): Batch indices, to only group
-            identical rows within batches. Defaults to None.
+        tensor: Tensor of shape [N, F].
+        batch: Batch indices, to only group identical rows within batches.
 
     Returns:
-        Tensor: List of group indices, from 0 to num. groups - 1, assigning all
+        List of group indices, from 0 to num. groups - 1, assigning all
             identical rows to the same group.
     """
     if batch is not None:
@@ -97,45 +110,50 @@ def group_by(data: Union[Data, Batch], keys: List[str]) -> LongTensor:
 
 
 def group_pulses_to_dom(data: Data) -> Data:
-    """Groups pulses on the same DOM, using DOM and string number."""
+    """Group pulses on the same DOM, using DOM and string number."""
     data.dom_index = group_by(data, ["dom_number", "string"])
     return data
 
 
 def group_pulses_to_pmt(data: Data) -> Data:
-    """Groups pulses on the same PMT, using PMT, DOM, and string number."""
+    """Group pulses on the same PMT, using PMT, DOM, and string number."""
     data.pmt_index = group_by(data, ["pmt_number", "dom_number", "string"])
     return data
 
 
-# Below mirroring `torch_geometric.nn.pool.{avg,max}_pool.py` exactly
-def _sum_pool_x(cluster, x, size: Optional[int] = None):
+# Below mirroring `torch_geometric.nn.pool.{avg,max}_pool.py`.
[…]
