### 2022-10-28 fd75092
Remove unused utils module
