### 2022-10-28 f9bcc8d
Add docstring
