### 2022-10-28 57ce6d0
Update detector.py
