+++ src/graphnet/models/detector/icecube.py
@@ -1,3 +1,5 @@
+"""IceCube-specific `Detector` class(es)."""
+
 import torch
 from torch_geometric.data import Data
 
@@ -17,16 +19,14 @@ class IceCube86(Detector):
     features = FEATURES.ICECUBE86
 
     def _forward(self, data: Data) -> Data:
-        """Ingests data, builds graph (connectivity/adjacency), and
-        preprocesses features.
+        """Ingest data, build graph, and preprocesses features.
 
         Args:
-            data (Data): Input graph data.
+            data: Input graph data.
 
         Returns:
-            Data: Connected and preprocessed graph data.
+            Connected and preprocessed graph data.
         """
-
         # Check(s)
         self._validate_features(data)
 
@@ -57,16 +57,14 @@ class IceCubeUpgrade(IceCubeDeepCore):
     features = FEATURES.UPGRADE
 
     def _forward(self, data: Data) -> Data:
-        """Ingests data, builds graph (connectivity/adjacency), and
-        preprocesses features.
+        """Ingest data, build graph, and preprocesses features.
 
         Args:
-            data (Data): Input graph data.
+            data: Input graph data.
 
         Returns:
-            Data: Connected and preprocessed graph data.
+            Connected and preprocessed graph data.
         """
-
         # Check(s)
         self._validate_features(data)
 
@@ -100,19 +98,18 @@ class IceCubeUpgrade_V2(IceCubeDeepCore):
 
     @property
     def nb_outputs(self):
+        """Return number of output features."""
         return self.nb_inputs + 3
 
     def _forward(self, data: Data) -> Data:
-        """Ingests data, builds graph (connectivity/adjacency), and
-        preprocesses features.
+        """Ingest data, build graph, and preprocesses features.
 
         Args:
-            data (Data): Input graph data.
+            data: Input graph data.
 
         Returns:
-            Data: Connected and preprocessed graph data.
+            Connected and preprocessed graph data.
         """
-
         # Check(s)
         self._validate_features(data)
 
