+++ src/graphnet/models/detector/__init__.py
@@ -1,3 +1,3 @@
-"""Detector-specific models, for data ingestion and standardisation."""
+"""Detector-specific modules, for data ingestion and standardisation."""
 
 from .icecube import IceCube86, IceCubeDeepCore
