+++ src/graphnet/models/task/task.py
@@ -1,5 +1,7 @@
+"""Base physics task-specific `Model` class(es)."""
+
 from abc import abstractmethod
-from typing import TYPE_CHECKING, List, Union
+from typing import TYPE_CHECKING, List, Tuple, Union
 from typing import Callable, Optional
 import numpy as np
 
@@ -9,7 +11,8 @@ try:
     from typing import final
 except ImportError:  # Python version < 3.8
 
-    def final(f):  # Identity decorator
+    # Identity decorator
+    def final(f):  # type: ignore  # noqa: D103
         return f
 
 
@@ -20,45 +23,18 @@ from torch_geometric.data import Data
 
 if TYPE_CHECKING:
     # Avoid cyclic dependency
-    from graphnet.training.loss_functions import LossFunction
+    from graphnet.training.loss_functions import LossFunction  # type: ignore[attr-defined]
 
 from graphnet.models import Model
 
 
 class Task(Model):
-    """Base class for all reconstruction and classification tasks.
-
-    Args:
-        hidden_size: The number of nodes in the layer feeding into this tasks,
-          used to construct the affine transformation to the predicted quantity.
-        target_label: Name of the quantity being predicted, used to extract the
-          target tensor from the `Data` object in `.compute_loss(...)`.
-        loss_function: Loss function appropriate to the task.
-        transform_prediction_and_target: Optional function to transform both the
-          predicted and target tensor before passing them to the loss function.
-          Useful e.g. for having the model predict quantities on a physical
-          scale, but transforming this scale to O(1) for a numerically stable
-          loss computation.
-        transform_target: Optional function to transform only the target tensor
-          before passing it, and the predicted tensor, to the loss function.
-          Useful e.g. for having the model predict a transformed version of the
-          target quantity, e.g. the log10-scaled energy, rather than the
-          physical quantity itself. Used in conjunction with `transform_inference`
-          to perform the inverse transform on the predicted quantity to recover
-          the physical scale.
-        transform_inference: Optional function to inverse-transform the model
-          prediction to recover a physical scale. Used in conjunction with
-          `transform_target`.
-        transform_support: Optional tuple to specify minimum and maximum of the
-          range of validity for the inverse transforms `transform_target` and
-          `transform_inference` in case this is restricted. By default the
-          invertibility of `transform_target` is tested on the range [-1e6, 1e6].
-    """
+    """Base class for all reconstruction and classification tasks."""
 
     @property
     @abstractmethod
     def nb_inputs(self) -> int:
-        """Number of inputs assumed by task."""
+        """Return number of inputs assumed by task."""
 
     @save_config
     def __init__(
@@ -69,10 +45,43 @@ class Task(Model):
         transform_prediction_and_target: Optional[Callable] = None,
         transform_target: Optional[Callable] = None,
         transform_inference: Optional[Callable] = None,
-        transform_support: Optional[tuple] = None,
+        transform_support: Optional[Tuple] = None,
         loss_weight: Optional[str] = None,
     ):
+        """Construct `Task`.
 
+        Args:
+            hidden_size: The number of nodes in the layer feeding into this
+                tasks, used to construct the affine transformation to the
+                predicted quantity.
+            target_labels: Name(s) of the quantity/-ies being predicted, used
+                to extract the  target tensor(s) from the `Data` object in
+                `.compute_loss(...)`.
+            loss_function: Loss function appropriate to the task.
+            transform_prediction_and_target: Optional function to transform
+                both the predicted and target tensor before passing them to the
+                loss function. Useful e.g. for having the model predict
+                quantities on a physical scale, but transforming this scale to
+                O(1) for a numerically stable loss computation.
+            transform_target: Optional function to transform only the target
[…]
