### 2022-10-28 f6b59fe
Remove unused PassOutput{1,2,3} classes
