### 2022-10-28 441527f
Move mypy configuration
