+++ setup.cfg
@@ -24,3 +24,4 @@ wrap-descriptions = 79
 
 [mypy]
 ignore_missing_imports = True
+follow-imports = silent
