### 2022-10-28 b2e9dbe
Move mypy configuration
