### 2022-10-28 9693e2e
Add code quality GitHub Action
