+++ .github/workflows/code-quality.yml
@@ -1,0 +1,54 @@
+# This is a simple workflow to check that all modified files satisfy code
+# quality requirements
+name: Code quality
+
+# Controls when the workflow will run
+on:
+  # Triggers the workflow on push and PRs
+  push:
+  pull_request:
+    branches: [ main ]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+
+  get-modified-files:
+    runs-on: ubuntu-latest
+    outputs:
+      all_changed_files: ${{ steps.changed-files.outputs.all_changed_files }}
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: Get changed files
+        id: changed-files
+        uses: tj-actions/changed-files@v34
+
+      - name: List all changed files
+        run: |
+          for file in ${{ steps.changed-files.outputs.all_changed_files }}; do
+            echo "$file was changed"
+          done
+
+      - name:
+        id: setvar
+        run: |
+          echo "::set-output name=all_files::${{ steps.changed-files.outputs.all_changed_files }};"
+
+  black:
+    needs: [ get-modified-files ]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install package
+        uses: ./.github/actions/install-cpu
+      - name: Run black
+        run: |
+          echo ${{needs.get-modified-files.outputs.all_files}}
+          black ${{needs.get-modified-files.outputs.all_files}} --check --diff --config black.toml
+
+
