+++ .github/workflows/code-quality.yml
@@ -15,10 +15,10 @@ on:
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
 
-  get-modified-files:
+  files:
     runs-on: ubuntu-latest
     outputs:
-      all_changed_files: ${{ steps.changed-files.outputs.all_changed_files }}
+      all: ${{ steps.setvar.outputs.all }}
     steps:
       - uses: actions/checkout@v3
         with:
@@ -26,21 +26,23 @@ jobs:
 
       - name: Get changed files
         id: changed-files
-        uses: tj-actions/changed-files@v34
+        uses: jitterbit/get-changed-files@v1
+        with:
+          format: ''
 
       - name: List all changed files
         run: |
-          for file in ${{ steps.changed-files.outputs.all_changed_files }}; do
+          for file in ${{ steps.changed-files.outputs.all }}; do
             echo "$file was changed"
           done
 
       - name:
         id: setvar
         run: |
-          echo "::set-output name=all_files::${{ steps.changed-files.outputs.all_changed_files }};"
+          echo "::set-output name=all::${{ steps.changed-files.outputs.all }};"
 
   black:
-    needs: [ get-modified-files ]
+    needs: [ files ]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
@@ -48,7 +50,7 @@ jobs:
         uses: ./.github/actions/install-cpu
       - name: Run black
         run: |
-          echo ${{needs.get-modified-files.outputs.all_files}}
-          black ${{needs.get-modified-files.outputs.all_files}} --check --diff --config black.toml
+          echo ${{needs.files.outputs.all}}
+          black ${{needs.files.outputs.all}} --check --diff --config black.toml
 
 
