+++ .github/workflows/code-quality.yml
@@ -42,6 +42,8 @@ jobs:
         run: |
           all_files=${{ steps.changed-files.outputs.all_changed_files }}
           python_files=$(grep -ow "[^ ]*\.py[^ ]*" <<<"$all_files")
+          echo "$all_files"
+          echo "$python_files"
           echo "all=$all_files" >> $GITHUB_OUTPUT
           echo "python=$python_files" >> $GITHUB_OUTPUT
 
