+++ .github/workflows/code-quality.yml
@@ -40,12 +40,8 @@ jobs:
       - name: Save to output
         id: setvar
         run: |
-          all_files=${{ steps.changed-files.outputs.all_changed_files }}
-          echo ">> $all_files"
-          python_files=$(grep -ow "[^ ]*\.py[^ ]*" <<<"$all_files")
-          echo ">> $python_files"
-          echo "all=$all_files" >> $GITHUB_OUTPUT
-          echo "python=$python_files" >> $GITHUB_OUTPUT
+          echo "all=${{ steps.changed-files.outputs.all_changed_files }}" >> $GITHUB_OUTPUT
+          echo "python=$(grep -ow '[^ ]*\.py[^ ]*'' <<<'${{ steps.changed-files.outputs.all_changed_files }}')" >> $GITHUB_OUTPUT
 
   black:
     needs: [ files ]
