+++ .github/workflows/code-quality.yml
@@ -41,7 +41,7 @@ jobs:
         id: setvar
         run: |
           echo "all=${{ steps.changed-files.outputs.all_changed_files }}" >> $GITHUB_OUTPUT
-          echo "python=$(grep -ow '[^ ]*\.py[^ ]*'' <<<'${{ steps.changed-files.outputs.all_changed_files }}')" >> $GITHUB_OUTPUT
+          echo "python=$(grep -ow '[^ ]*\.py[^ ]*' <<<'${{ steps.changed-files.outputs.all_changed_files }}')" >> $GITHUB_OUTPUT
 
   black:
     needs: [ files ]
