+++ .github/workflows/code-quality.yml
@@ -21,15 +21,13 @@ jobs:
       all: ${{ steps.setvar.outputs.all }}
       python: ${{ steps.setvar.outputs.python }}
     steps:
-      - uses: actions/checkout@v2
-        #with:
-        #  fetch-depth: 0
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
 
       - name: Get changed files
         id: changed-files
         uses: tj-actions/changed-files@v34
-        with:
-          since_last_remote_commit: 'true'
 
       - name: List all changed files
         run: |
@@ -52,7 +50,8 @@ jobs:
         uses: ./.github/actions/install-cpu
       - name: Run black
         run: |
-          echo ${{needs.files.outputs.python}}
-          black ${{needs.files.outputs.python}} --check --diff --config black.toml
+          echo "${{ needs.files.outputs.all }}"
+          echo "${{ needs.files.outputs.python }}"
+          black ${{ needs.files.outputs.python }} --check --diff --config black.toml
 
 
