+++ .github/workflows/code-quality.yml
@@ -39,7 +39,8 @@ jobs:
         id: setvar
         run: |
           all_files="${{ steps.changed-files.outputs.all_changed_files }}"
-          python_files="$(grep -ow '[^ ]*\.py[^ ]*' <<<'${{ steps.changed-files.outputs.all_changed_files }}')"
+          python_files="$(grep -ow '[^ ]*\.py[^ ]*' <<<'$all_files')"
+
           echo all="$all_files" >> $GITHUB_OUTPUT
           echo python="$python_files" >> $GITHUB_OUTPUT
 
