+++ .github/workflows/code-quality.yml
@@ -29,12 +29,6 @@ jobs:
         id: changed-files
         uses: tj-actions/changed-files@v34
 
-      - name: List all changed files
-        run: |
-          for file in ${{ steps.changed-files.outputs.all_changed_files }}; do
-            echo "$file was changed"
-          done
-
       - name: Save to output
         id: setvar
         run: |
@@ -53,8 +47,48 @@ jobs:
         uses: ./.github/actions/install-cpu
       - name: Run black
         run: |
-          echo "${{ needs.files.outputs.all }}"
-          echo "${{ needs.files.outputs.python }}"
-          black ${{ needs.files.outputs.python }} --check --diff --config black.toml
+          pre-commit run black --files ${{ needs.files.outputs.python }}
 
+  flake8:
+    needs: [ files ]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install package
+        uses: ./.github/actions/install-cpu
+      - name: Run flake8
+        run: |
+          pre-commit run flake8 --files ${{ needs.files.outputs.python }}
 
+  docformatter:
+    needs: [ files ]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install package
+        uses: ./.github/actions/install-cpu
+      - name: Run docformatter
+        run: |
+          pre-commit run docformatter --files ${{ needs.files.outputs.python }}
+
+  pydocstyle:
+    needs: [ files ]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install package
+        uses: ./.github/actions/install-cpu
+      - name: Run pydocstyle
+        run: |
+          pre-commit run pydocstyle --files ${{ needs.files.outputs.python }}
+
+  mypy:
+    needs: [ files ]
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install package
+        uses: ./.github/actions/install-cpu
+      - name: Run mypy
+        run: |
+          pre-commit run mypy --files ${{ needs.files.outputs.python }}
