+++ .github/workflows/code-quality.yml
@@ -42,7 +42,7 @@ jobs:
     needs: [ files ]
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Install package
         uses: ./.github/actions/install-cpu
       - name: Run black
@@ -53,7 +53,7 @@ jobs:
     needs: [ files ]
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Install package
         uses: ./.github/actions/install-cpu
       - name: Run flake8
@@ -64,7 +64,7 @@ jobs:
     needs: [ files ]
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Install package
         uses: ./.github/actions/install-cpu
       - name: Run docformatter
@@ -75,7 +75,7 @@ jobs:
     needs: [ files ]
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Install package
         uses: ./.github/actions/install-cpu
       - name: Run pydocstyle
@@ -86,7 +86,7 @@ jobs:
     needs: [ files ]
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Install package
         uses: ./.github/actions/install-cpu
       - name: Run mypy
