### 2022-10-28 04a6cd4
Update src/graphnet/training/loss_functions.py
device inheritance

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
