+++ src/graphnet/training/loss_functions.py
@@ -120,7 +120,7 @@ class MultiClassificationCrossEntropyLoss(LossFunction):
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
         # transformer for multiple pid types to total of three classes
-        device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        device = prediction.device
         pid_transform = {1:0, 12:2, 13:1, 14:2, 16:2}
         target_new = one_hot(torch.tensor([pid_transform[np.abs(int(value))] for value in target]), 3).to(device)
         return cross_entropy(
