### 2022-10-28 afecd26
Update src/graphnet/training/loss_functions.py
loss function renaming

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
