### 2022-10-28 b4725f4
Update src/graphnet/training/loss_functions.py
description reformulation

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
