### 2022-10-28 4384bbd
Update src/graphnet/training/loss_functions.py
trimming description

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
