### 2022-10-28 5ed1076
Update src/graphnet/training/loss_functions.py
description reformulation

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
