+++ src/graphnet/training/loss_functions.py
@@ -129,9 +129,10 @@ class CrossEntropyLoss(LossFunction):
         )
 
 class BinaryCrossEntropyLoss(LossFunction):
-    """Computes binary cross entropy for a vector of predictions (between 0 and 1),
-    targets should be 0 and 1 for muon and neutrino respectively
-    where prediction is prob. the PID is neutrino (12,14,16)
+    """Compute binary cross entropy loss.
+    
+    Predictions are vector probabilities (i.e., values between 0 and 1), and 
+    targets should be 0 and 1.
     """
 
     def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
