### 2022-10-28 36ea9e9
Update src/graphnet/models/task/reconstruction.py
dynamical multiclass classification

Co-authored-by: Andreas Søgaard <andreas.sogaard@gmail.com>
