+++ src/graphnet/models/task/reconstruction.py
@@ -167,13 +167,21 @@ class TimeReconstruction(Task):
         # Leave as it is
         return x
 
-class MulticlassificationTask(Task):
-    # requires three features: probability of being noise, muon or neutrino.
-    nb_inputs = 3
+class MulticlassClassificationTask(Task):
+    # Requires the same number of features as the number of classes being predicted
+    @property
+    def nb_inputs(self):
+        return self._nb_inputs
+
+    @save_config
+    def __init__(self, nb_inputs, *args, **kwargs):
+        self._nb_inputs = nb_inputs
+        self._softmax = torch.nn.Softmax()
+        super().__init__(self, *args, **kwargs)
 
     def _forward(self, x):
-        # transform probability of being noise, muon or neutrino
-        return torch.sigmoid(x)
+        # Transform latent features into probabilities.
+        return self._softmax(x)
 
 class BinaryClassificationTask(Task):
     # requires one feature: probability of being neutrino?
