+++ .github/workflows/code-quality.yml
@@ -16,6 +16,7 @@ on:
 jobs:
 
   files:
+    name: Get changed files
     runs-on: ubuntu-latest
     outputs:
       all: ${{ steps.setvar.outputs.all }}
@@ -39,6 +40,7 @@ jobs:
           echo python="$python_files" >> $GITHUB_OUTPUT
 
   pre-commit:
+    name: Pre-commit - ${{ matrix.hook }}
     needs: [ files ]
     runs-on: ubuntu-latest
     strategy:
