+++ .github/workflows/paper.yml
@@ -1,5 +1,5 @@
 # This is a simple action to generate paper draft PDFs.
-name: paper
+name: Paper
 
 on:
   push:
@@ -10,6 +10,7 @@ on:
 
 jobs:
   paper-2022-joss:
+    name: Compile
     runs-on: ubuntu-latest
     name: paper-draft-pdf
     steps:
