+++ .github/workflows/paper.yml
@@ -10,9 +10,8 @@ on:
 
 jobs:
   paper-2022-joss:
-    name: Compile
+    name: Compile draft PDF
     runs-on: ubuntu-latest
-    name: paper-draft-pdf
     steps:
       - name: Checkout
         uses: actions/checkout@v2
