### 2022-10-29 07f9edf
Update pre-commit configuration
