+++ setup.cfg
@@ -24,4 +24,3 @@ wrap-descriptions = 79
 
 [mypy]
 ignore_missing_imports = True
-follow-imports = silent
