### 2022-10-29 10c7179
Update pre-commit configuration
