### 2022-10-29 9a9f183
Update sqlite_dataconverter.py
