+++ src/graphnet/data/sqlite/sqlite_dataconverter.py
@@ -1,24 +1,26 @@
-from collections import OrderedDict, Iterable
-import json
-import numpy as np
+"""DataConverter for the SQLite backend."""
+
+from collections import OrderedDict
 import os
+from typing import Any, Dict, List, Optional, Tuple, Union
+
 import pandas as pd
 import sqlalchemy
 import sqlite3
 from tqdm import tqdm
-from typing import Any, Dict, List, MutableMapping, Optional, Tuple, Union
 
-from graphnet.data.dataconverter import DataConverter
+from graphnet.data.dataconverter import DataConverter  # type: ignore[attr-defined]
 from graphnet.data.sqlite.sqlite_utilities import run_sql_code, save_to_sql
 
 
 class SQLiteDataConverter(DataConverter):
+    """Class for converting I3-files to SQLite format."""
 
     # Class variables
     file_suffix = "db"
 
     # Abstract method implementation(s)
-    def save_data(self, data: List[OrderedDict], output_file: str):
+    def save_data(self, data: List[Dict[str, Any]], output_file: str):
         """Save data to SQLite database."""
         # Check(s)
         if os.path.exists(output_file):
@@ -27,10 +29,10 @@ class SQLiteDataConverter(DataConverter):
             )
 
         # Concatenate data
-        dataframe = OrderedDict()
+        dataframe: Dict[str, pd.DataFrame] = OrderedDict()
         for data_dict in data:
-            for key, data in data_dict.items():
-                df = construct_dataframe(data)
+            for key, data_values in data_dict.items():
+                df = construct_dataframe(data_values)
 
                 if self.any_pulsemap_is_non_empty(data_dict) and len(df) > 0:
                     # only include data_dict in temp. databases if at least one pulsemap is non-empty,
@@ -58,6 +60,14 @@ class SQLiteDataConverter(DataConverter):
     def merge_files(
         self, output_file: str, input_files: Optional[List[str]] = None
     ):
+        """SQLite-specific method for merging output files/databases.
+
+        Args:
+            output_file: Name of the output file containing the merged results.
+            input_files: Intermediate files/databases to be merged, according
+                to the specific implementation. Default to None, meaning that
+                all files/databases output by the current instance are merged.
+        """
         if input_files is None:
             self.info("Merging files output by current instance.")
             input_files = self._output_files
@@ -94,7 +104,9 @@ class SQLiteDataConverter(DataConverter):
             self.warning("No temporary database files found!")
 
     # Internal methods
-    def _extract_table_names(self, db: Union[str, List[str]]) -> Tuple[str]:
+    def _extract_table_names(
+        self, db: Union[str, List[str]]
+    ) -> Tuple[str, ...]:
         """Get the names of all tables in database `db`."""
         if isinstance(db, list):
             results = [self._extract_table_names(path) for path in db]
@@ -125,7 +137,7 @@ class SQLiteDataConverter(DataConverter):
                 return columns
         return []
 
-    def any_pulsemap_is_non_empty(self, data_dict: OrderedDict) -> bool:
+    def any_pulsemap_is_non_empty(self, data_dict: Dict[str, Dict]) -> bool:
         """Check whether there are non-empty pulsemaps extracted from P frame.
 
         Takes in the data extracted from the P frame, then retrieves the
@@ -138,11 +150,11 @@ class SQLiteDataConverter(DataConverter):
         if len(self._pulsemaps) == 0:
             return True
 
-        pulsemap_dicts = map(data_dict.get, self._pulsemaps)
+        pulsemap_dicts = [data_dict[pulsemap] for pulsemap in self._pulsemaps]
         return any(d["dom_x"] for d in pulsemap_dicts)
 
     def _attach_index(self, database: str, table_name: str):
-        """Attaches the table index.
+        """Attach the table index.
 
         Important for query times!
[…]
