### 2022-10-29 30d1922
Update sqlite_selection.py
