### 2022-10-31 3ab9f9f
move dependencies to import and line reductions
