### 2022-10-31 ac50af0
added all functions to main as testables
