### 2022-10-31 5390441
setup for testing all test functions in script
