### 2022-11-01 a3e94e7
Update pre-commit config to enforce type hinting with mypy
