### 2022-11-01 c5e79d6
Update parquet_to_sqlite.py
