+++ src/graphnet/deployment/i3modules/graphnet_module.py
@@ -1,7 +1,7 @@
-"""Classes inheriting from I3Module, for use in deploying GNN models to
-icetray."""
+"""Class(es) for deploying GraphNeT models in icetray as I3Modules."""
 
 import os.path
+from typing import TYPE_CHECKING, Any, List, Union
 
 import numpy as np
 import torch
@@ -14,12 +14,10 @@ from graphnet.data.extractors import (
 )
 from graphnet.data.constants import FEATURES
 from graphnet.models import StandardModel
+from graphnet.models.model import Model
 from graphnet.utilities.imports import has_icecube_package
-from graphnet.utilities.logging import get_logger
 
-logger = get_logger()
-
-if has_icecube_package():
+if has_icecube_package() or TYPE_CHECKING:
     from icecube.icetray import (
         I3Module,
         I3Frame,
@@ -33,16 +31,16 @@ class GraphNeTModuleBase(I3Module):
     """Base I3Module for running graphnet models in I3Tray chains."""
 
     # Class variables
-    FEATURES = None
-    I3FEATUREEXTRACTOR_CLASS = None
+    FEATURES: List[str]
+    I3FEATUREEXTRACTOR_CLASS: type
     DTYPES = {
         "float16": torch.float16,
         "float32": torch.float32,
         "float64": torch.float64,
     }
 
-    def __init__(self, context):
-
+    def __init__(self, context: Any) -> None:
+        """Construct `GraphNeTModuleBase`."""
         # Check
         if self.FEATURES is None:
             raise Exception("Please use an experiment-specific I3Module.")
@@ -58,19 +56,18 @@ class GraphNeTModuleBase(I3Module):
         self.AddParameter("dtype", "doc_string__dtype", "float32")
 
         # Standard member variables
-        self.keys = None
-        self.model = None
-        self.dtype = None
+        self.keys: Union[str, List[str]]
+        self.model: Model
+        self.dtype: torch.dtype
 
-    def Configure(self):  # pylint: disable=invalid-name
+    def Configure(self) -> None:  # pylint: disable=invalid-name
         """Configure I3Module based on keyword parameters."""
-
         # Extract parameters
-        keys = self.GetParameter("keys")
-        gcd_file = self.GetParameter("gcd_file")
-        model = self.GetParameter("model")
-        pulsemaps = self.GetParameter("pulsemaps")
-        dtype = self.GetParameter("dtype")
+        keys: Union[str, List[str]] = self.GetParameter("keys")
+        gcd_file: str = self.GetParameter("gcd_file")
+        model: Union[str, Model] = self.GetParameter("model")
+        pulsemaps: Union[str, List[str]] = self.GetParameter("pulsemaps")
+        dtype: str = self.GetParameter("dtype")
 
         # Check(s)
         assert keys is not None
@@ -104,11 +101,12 @@ class GraphNeTModuleBase(I3Module):
         # predictions are applied.
         self.model.inference()
 
-    def Physics(self, frame: I3Frame):  # pylint: disable=invalid-name
+    def Physics(
+        self, frame: I3Frame
+    ) -> None:  # py-l-i-n-t-:- -d-i-s-able=invalid-name
         """Process Physics I3Frame and write predictions."""
-
         # Extract features
-        features = self.extract_feature_array_from_frame(frame)
+        features = self._extract_feature_array_from_frame(frame)
 
         # Prepare graph data
         n_pulses = torch.tensor([features.shape[0]], dtype=torch.int32)
@@ -133,14 +131,14 @@ class GraphNeTModuleBase(I3Module):
                 predictions
[…]
