### 2022-11-01 7cdfc57
Update dynedge.py
