+++ src/graphnet/models/gnn/dynedge.py
@@ -1,5 +1,5 @@
 """Implementation of the DynEdge GNN model architecture."""
-from typing import List, Optional, Tuple, Union
+from typing import List, Optional, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor, LongTensor
@@ -20,54 +20,53 @@ GLOBAL_POOLINGS = {
 
 
 class DynEdge(GNN):
+    """DynEdge (dynamical edge convolutional) model."""
+
     @save_config
     def __init__(
         self,
         nb_inputs: int,
         *,
-        nb_neighbours: Optional[int] = 8,
-        features_subset: Optional[List[int]] = None,
-        dynedge_layer_sizes: Optional[List[Tuple[int]]] = None,
+        nb_neighbours: int = 8,
+        features_subset: Optional[Union[List[int], slice]] = None,
+        dynedge_layer_sizes: Optional[List[Tuple[int, ...]]] = None,
         post_processing_layer_sizes: Optional[List[int]] = None,
         readout_layer_sizes: Optional[List[int]] = None,
         global_pooling_schemes: Optional[Union[str, List[str]]] = None,
         add_global_variables_after_pooling: bool = False,
     ):
-        """DynEdge (dynamical edge convolutional) model.
+        """Construct `DynEdge`.
 
         Args:
-            nb_inputs (int): Number of input features on each node
-            nb_neighbours (Optional[int], optional): Number of neighbours to
-                used in the k-nearest neighbour clustering which is performed
-                after each (dynamical) edge convolution. Defaults to 8.
-            features_subset (Optional[List[int]], optional): The subset of
-                latent features on each node that are used as metric dimensions
-                when performing the k-nearest neighbours clustering. Defaults
-                to slice(0,3).
-            dynedge_layer_sizes (Optional[List[Tuple[int]]], optional): The
-                layer sizes, or latent feature dimenions, used in the
-                `DynEdgeConv` layer. Each entry in `dynedge_layer_sizes`
-                corresponds to a single `DynEdgeConv` layer; the integers in
-                the corresponding tuple corresponds to the layer sizes in the
-                multi-layer perceptron (MLP) that is applied within each
-                `DynEdgeConv` layer. That is, a list of size-two tuples means
-                that all `DynEdgeConv` layers contain a two-layer MLP.
+            nb_inputs: Number of input features on each node.
+            nb_neighbours: Number of neighbours to used in the k-nearest
+                neighbour clustering which is performed after each (dynamical)
+                edge convolution.
+            features_subset: The subset of latent features on each node that
+                are used as metric dimensions when performing the k-nearest
+                neighbours clustering. Defaults to [0,1,2].
+            dynedge_layer_sizes: The layer sizes, or latent feature dimenions,
+                used in the `DynEdgeConv` layer. Each entry in
+                `dynedge_layer_sizes` corresponds to a single `DynEdgeConv`
+                layer; the integers in the corresponding tuple corresponds to
+                the layer sizes in the multi-layer perceptron (MLP) that is
+                applied within each `DynEdgeConv` layer. That is, a list of
+                size-two tuples means that all `DynEdgeConv` layers contain a
+                two-layer MLP.
                 Defaults to [(128, 256), (336, 256), (336, 256), (336, 256)].
-            post_processing_layer_sizes (Optional[List[int]], optional): Hidden
-                layer sizes in the MLP following the skip-concatenation of the
-                outputs of each `DynEdgeConv` layer. Defaults to [336, 256].
-            readout_layer_sizes (Optional[List[int]], optional): Hidden layer
-                sizes in the MLP following the post-processing _and_ optional
-                global pooling. As this is the last layer(s) in the model, the
-                last layer in the read-out yields the output of the `DynEdge`
[…]
