+++ src/graphnet/utilities/filesys.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 import re
-from typing import List, Optional, Union, cast
+from typing import List, Optional, Tuple, Union, cast
 
 
 def is_gcd_file(filename: str) -> bool:
@@ -32,21 +32,21 @@ def has_extension(filename: str, extensions: List[str]) -> bool:
 
 def find_i3_files(
     directories: Union[str, List[str]], gcd_rescue: Optional[str] = None
-):
+) -> Tuple[List[str], List[str]]:
     """Find I3 files and corresponding GCD files in `directories`.
 
-    Finds I3 files in dir and matches each file with a corresponding GCD file if
-    present in the directory, matches with gcd_rescue if gcd is not present in
-    the directory.
+    Finds I3 files in dir and matches each file with a corresponding GCD file
+    if present in the directory, matches with gcd_rescue if gcd is not present
+    in the directory.
 
     Args:
-        directories (list[str]): Directories to search recursively for I3 files.
-        gcd_rescue (str): Path to the GCD that will be default if no GCD is
-            present in the directory.
+        directories: Directories to search recursively for I3 files.
+        gcd_rescue: Path to the GCD that will be default if no GCD is present
+            in the directory.
 
     Returns:
-        i3_list (list[str]): Paths to I3 files in `directories`
-        gcd_list (list[str]): Paths to GCD files for each I3 file.
+        i3_list: Paths to I3 files in `directories`
+        gcd_list: Paths to GCD files for each I3 file.
     """
     if isinstance(directories, str):
         directories = [directories]
@@ -72,13 +72,12 @@ def find_i3_files(
                 str(path) for path in paths if path.parent == folder
             ]
             folder_i3_files = list(filter(is_i3_file, folder_files))
-            folder_gcd_files = cast(
-                List[Optional[str]], list(filter(is_gcd_file, folder_files))
-            )
+            folder_gcd_files = list(filter(is_gcd_file, folder_files))
 
             # Make sure that no more than one GCD file is found; and use rescue file of none is found.
             assert len(folder_gcd_files) <= 1
             if len(folder_gcd_files) == 0:
+                assert gcd_rescue is not None
                 folder_gcd_files = [gcd_rescue]
 
             # Store list of I3 files and corresponding GCD files.
