+++ src/graphnet/utilities/imports.py
@@ -1,6 +1,7 @@
 """Common functions for icetray/data-based unit tests."""
 
 from functools import wraps
+from typing import Any, Callable
 
 from graphnet.utilities.logging import get_logger, warn_once
 
@@ -35,11 +36,11 @@ def has_torch_package() -> bool:
         return False
 
 
-def requires_icecube(test_function):
+def requires_icecube(test_function: Callable) -> Callable:
     """Decorate `test_function` for use only if `icecube` module is present."""
 
     @wraps(test_function)
-    def wrapper(*args, **kwargs):
+    def wrapper(*args: Any, **kwargs: Any) -> Any:
         if has_icecube_package():
             return test_function(*args, **kwargs)
         else:
