### 2022-11-02 6c188a8
Create citation.cff
Based on the documentation from [here](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files), I have made a draft of a potential citation file with names of contributors and their corresponding orcid.
Feel free to give any and all input and thoughts about it :)
