+++ citation.cff
@@ -1,0 +1,39 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Søgaard"
+  given-names: "Andreas"
+  orcid: "https://orcid.org/0000-0002-0823-056X"
+- family-names: "C. Petersen"
+  given_names: "Troels"
+  orcid: "https://orcid.org/0000-0003-0221-3037"
+- family-names: "F. Ørsøe"
+  given-names: "Rasmus"
+  orcid: "https://orcid.org/0000-0001-8890-4124"
+- family-names: "Eller"
+  given_names: "Philipp"
+  orcid: "https://orcid.org/0000-0001-6354-5209"
+- family-names: "Ha Minh"
+  given-names: "Martin"
+  orcid: "https://orcid.org/0000-0001-7776-4875"
+- family-names: "Bozianu"
+  given-names: "Leon"
+  orcid: "https://orcid.org/0000-0002-1243-9980"
+- family-names: "E. Iversen"
+  given-names: "Kaare"
+  orcid: "https://orcid.org/0000-0001-6533-4085"
+- family-names: "Holm"
+  given-names: "Morten"
+  orcid: "https://orcid.org/0000-0003-1383-2810"
+- family-names: "Guggenmos"
+  given-names: "Tim"
+  orcid: ""
+- family-names: "Andresen"
+  given-names: "Peter"
+  orcid: ""
+
+title: "GraphNet"
+version: 0.2.2
+doi: 
+date-released: 2022-11-02
+url: "https://github.com/graphnet-team/graphnet"
