+++ src/graphnet/data/extractors/__init__.py
@@ -6,3 +6,4 @@ from .i3splinempeextractor import I3SplineMPEICExtractor
 from .i3tumextractor import I3TUMExtractor
 from .i3hybridrecoextractor import I3GalacticPlaneHybridRecoExtractor
 from .i3genericextractor import I3GenericExtractor
+from .i3pisaextractor import I3PISAExtractor
