+++ _sources/graphnet.data.extractors.rst.txt
@@ -44,6 +44,14 @@ graphnet.data.extractors.i3hybridrecoextractor module
    :undoc-members:
    :show-inheritance:
 
+graphnet.data.extractors.i3pisaextractor module
+-----------------------------------------------
+
+.. automodule:: graphnet.data.extractors.i3pisaextractor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 graphnet.data.extractors.i3retroextractor module
 ------------------------------------------------
 
