### 2022-11-03 cb4423d
Update docstring
