### 2022-11-03 4311216
Remove the deprecated 'set-output' command in GitHub Actions
