### 2022-11-03 7e0eb63
Update examples/bjoern_low_weight_fitter_example.py
