+++ examples/bjoern_low_weight_fitter_example.py
@@ -1,4 +1,5 @@
-import sqlite3
+"""Example of how to use WeightFitter class, specifically BjoernLow."""
+
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
