### 2022-11-03 98342a6
Update examples/bjoern_low_weight_fitter_example.py
