+++ examples/bjoern_low_weight_fitter_example.py
@@ -1,31 +1,37 @@
-"""Example of how to use WeightFitter class, specifically BjoernLow."""
+"""Example of using the WeightFitter class, specifically BjoernLow."""
 
 import numpy as np
 import matplotlib.pyplot as plt
 
 from graphnet.training.weight_fitting import BjoernLow
 
-database = "/my_databases/my_database/data/my_database.db"
 
-# Choose truth variable
-variable = "energy"
+def main() -> None:
+    """Run example."""
+    # Configuration
+    database = "/my_databases/my_database/data/my_database.db"
+    variable = "energy"
+    bins = bins = np.arange(0, 5, 0.01)  # in log10
 
-# Choose binning
-bins = bins = np.arange(0, 5, 0.01)  # in log10
+    # Fit the uniform weights
+    fitter = BjoernLow(database)
+    weights = fitter.fit(
+        bins=bins,
+        variable=variable,
+        add_to_database=True,
+        transform=np.log10,
+        x_low=1.5,
+    )
 
-# Fit the uniform weights
-fitter = BjoernLow(database)
-weights = fitter.fit(
-    bins=bins,
-    variable="energy",
-    add_to_database=True,
-    transform=np.log10,
-    x_low=1.5,
-)
+    # Plot the results
+    fig = plt.figure()
+    plt.hist(
+        weights["energy"],
+        bins=bins,
+        weights=weights["energy_bjørn_low_weight"],
+    )
+    fig.savefig("test_hist.png")
 
-# Plot the results
-fig = plt.figure()
-plt.hist(
-    weights["energy"], bins=bins, weights=weights["energy_bjørn_low_weight"]
-)
-fig.savefig("test_hist.png")
+
+if __name__ == "__main__":
+    main()
