### 2022-11-03 2a97556
Update examples/contour_fitter_example.py
