### 2022-11-03 8370c26
Update examples/train_direction_reco.py
