### 2022-11-03 8691beb
Update examples/train_direction_reco.py
