### 2022-11-03 272ae6d
Update examples/compare_sqlite_and_parquet.py
