### 2022-11-03 5e898ad
Update examples/train_direction_reco.py
