### 2022-11-03 ddbb6bf
Update examples/train_model.py
