+++ examples/plot_feature_distributions.py
@@ -1,3 +1,5 @@
+"""Example of plotting feature distributions from SQLite database."""
+
 import matplotlib.pyplot as plt
 import numpy as np
 from tqdm import tqdm
@@ -19,8 +21,9 @@ features = FEATURES.UPGRADE
 truth = TRUTH.UPGRADE
 
 
-# Main function definition
-def main():
+def main() -> None:
+    """Run example."""
+    # Remove `interaction_time` if it exists
     try:
         del truth[truth.index("interaction_time")]
     except ValueError:
@@ -63,14 +66,14 @@ def main():
     )
 
     # Get feature matrix
-    x_original = []
-    x_preprocessed = []
+    x_original_list = []
+    x_preprocessed_list = []
     for batch in tqdm(training_dataloader):
-        x_original.append(batch.x.numpy())
-        x_preprocessed.append(detector(batch).x.numpy())
+        x_original_list.append(batch.x.numpy())
+        x_preprocessed_list.append(detector(batch).x.numpy())
 
-    x_original = np.concatenate(x_original, axis=0)
-    x_preprocessed = np.concatenate(x_preprocessed, axis=0)
+    x_original = np.concatenate(x_original_list, axis=0)
+    x_preprocessed = np.concatenate(x_preprocessed_list, axis=0)
 
     logger.info("Number of NaNs:", np.sum(np.isnan(x_original)))
     logger.info("Number of infs:", np.sum(np.isinf(x_original)))
@@ -111,6 +114,5 @@ def main():
     fig.savefig("feature_distribution_preprocessed.png")
 
 
-# Main function call
 if __name__ == "__main__":
     main()
