### 2022-11-03 17c3ef6
Update examples/contour_plotting_example.py
