### 2022-11-03 b75dcfc
Update examples/uniform_weight_fitter_example.py
