### 2022-11-03 6953b16
Update examples/pisa_weight_fitter_example.py
