### 2022-11-03 e260381
Update examples/convert_i3_files.py
