+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Unit test suite for the code in `src/graphnet`."""
