### 2022-11-03 cef3304
Update docstring
