+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Unit test suite for the code in `src/graphnet`."""
+"""Unit test suite for the code in `graphnet`."""
