### 2022-11-03 3c3fc25
Remove duplicate unit test
