+++ tests/data/test_dataconverters_and_datasets.py
@@ -96,77 +96,6 @@ def test_dataconverter(
     assert os.path.exists(path), path
 
 
-def test_i3genericextractor(test_data_dir: str = TEST_DATA_DIR) -> None:
-    """Test the implementation of `I3GenericExtractor`."""
-    # Constants(s)
-    mc_tree = "I3MCTree"
-    pulse_series = "SRTInIcePulses"
-
-    # Constructor I3Extractor instance(s)
-    generic_extractor = I3GenericExtractor(keys=[mc_tree, pulse_series])
-    truth_extractor = I3TruthExtractor()
-    feature_extractor = I3FeatureExtractorIceCube86(pulse_series)
-
-    i3_file = os.path.join(test_data_dir, FILE_NAME) + ".i3.zst"
-    gcd_file = os.path.join(test_data_dir, GCD_FILE)
-
-    generic_extractor.set_files(i3_file, gcd_file)
-    truth_extractor.set_files(i3_file, gcd_file)
-    feature_extractor.set_files(i3_file, gcd_file)
-
-    i3_file_io = dataio.I3File(i3_file, "r")
-    ix_test = 10
-    while i3_file_io.more():
-        try:
-            frame = i3_file_io.pop_physics()
-        except:  # noqa: E722
-            continue
-
-        generic_data = generic_extractor(frame)
-        truth_data = truth_extractor(frame)
-        feature_data = feature_extractor(frame)
-
-        if ix_test == 10:
-            print(list(generic_data[pulse_series].keys()))
-            print(list(truth_data.keys()))
-            print(list(feature_data.keys()))
-
-        # Truth vs. generic
-        key_pairs = [
-            ("energy", "energy"),
-            ("zenith", "dir__zenith"),
-            ("azimuth", "dir__azimuth"),
-            ("pid", "pdg_encoding"),
-        ]
-
-        for truth_key, generic_key in key_pairs:
-            assert (
-                truth_data[truth_key]
-                == generic_data[f"{mc_tree}__primaries"][generic_key][0]
-            )
-
-        # Reco vs. generic
-        key_pairs = [
-            ("charge", "charge"),
-            ("dom_time", "time"),
-            ("dom_x", "position__x"),
-            ("dom_y", "position__y"),
-            ("dom_z", "position__z"),
-            ("width", "width"),
-            ("pmt_area", "area"),
-            # ("rde", "relative_dom_efficiency"),  <-- Missing
-        ]
-
-        for reco_key, generic_key in key_pairs:
-            assert np.allclose(
-                feature_data[reco_key], generic_data[pulse_series][generic_key]
-            )
-
-        ix_test -= 1
-        if ix_test == 0:
-            break
-
-
 @pytest.mark.order(3)
 @pytest.mark.parametrize("backend", ["sqlite", "parquet"])
 def test_dataset(backend: str, test_data_dir: str = TEST_DATA_DIR) -> None:
@@ -230,5 +159,4 @@ def test_dataset(backend: str, test_data_dir: str = TEST_DATA_DIR) -> None:
 
 
 if __name__ == "__main__":
-    # test_dataconverter("sqlite")
-    test_i3genericextractor()
+    test_dataconverter("sqlite")
