### 2022-11-03 bf528e4
Update tests/data/test_i3extractor.py
