+++ tests/data/test_i3extractor.py
@@ -7,19 +7,19 @@ from graphnet.data.extractors import (
 )
 
 
-def test_featureextractor_constructor():
+def test_featureextractor_constructor() -> None:
     """Test that the default constructor works."""
     extractor = I3FeatureExtractorIceCube86("pulsemap")
     assert extractor is not None
 
 
-def test_truthextractor_constructor():
+def test_truthextractor_constructor() -> None:
     """Test that the default constructor works."""
     extractor = I3TruthExtractor()
     assert extractor is not None
 
 
-def test_retroextractor_constructor():
+def test_retroextractor_constructor() -> None:
     """Test that the default constructor works."""
     extractor = I3RetroExtractor()
     assert extractor is not None
