### 2022-11-03 cb1d5da
Update tests/models/test_coarsening.py
