+++ tests/models/test_task.py
@@ -1,3 +1,5 @@
+"""Unit tests for Task classes."""
+
 import pytest
 import torch
 
@@ -8,7 +10,8 @@ from graphnet.models.task.reconstruction import EnergyReconstruction
 from graphnet.training.loss_functions import LogCoshLoss
 
 
-def test_transform_prediction_and_target():
+def test_transform_prediction_and_target() -> None:
+    """Test implementation of `transform_*` arguments to `Task`."""
     detector = IceCube86(
         graph_builder=KNNGraphBuilder(nb_nearest_neighbours=8),
     )
@@ -19,7 +22,11 @@ def test_transform_prediction_and_target():
     # Test not inverse functions
     with pytest.raises(
         AssertionError,
-        match="The provided transforms for targets during training and predictions during inference are not inverse. Please adjust transformation functions or support.",
+        match=(
+            "The provided transforms for targets during training and "
+            "predictions during inference are not inverse. Please adjust "
+            "transformation functions or support."
+        ),
     ):
         EnergyReconstruction(
             hidden_size=gnn.nb_outputs,
@@ -32,7 +39,9 @@ def test_transform_prediction_and_target():
     # Test wrong combination of inputs
     with pytest.raises(
         AssertionError,
-        match="Please specify both `transform_inference` and `transform_target`",
+        match=(
+            "Please specify both `transform_inference` and `transform_target`"
+        ),
     ):
         EnergyReconstruction(
             hidden_size=gnn.nb_outputs,
