### 2022-11-03 a29e2a2
Update tests/models/test_config.py
