### 2022-11-08 #4831 @JostMigenda
Sure, sign me up! I’ll just note right up front that I have little experience with GNNs; so it might be good if the second reviewer is more experienced in that area.
