### 2022-11-26 #4971 @dfm
@JostMigenda, @GageDeZoort — This is the review thread for the paper. All of our correspondence will happen here from now on. Thanks again for agreeing to participate!

We aim for the review process to be completed within about 4-6 weeks but please try to make a start ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule.
