### 2022-12-13 graphnet-team/graphnet#373 @JostMigenda
Installation fails on macOS
**Describe the bug**
As part of openjournals/joss-reviews#4971 I’m trying to install GraphNeT on my local machine (macOS 12.6.1). Since I’m not an IceCube collaborator and thus don’t have access to IceTray, I’m using the stand-alone installation described in the README file. This fails because the requirement `

**To Reproduce**
Steps to reproduce the behavior:
1. Start clean shell.
2. Run the following commands:
```bash
$ git clone git@github.com:graphnet-team/graphnet.git
$ cd graphnet
$ conda create --name graphnet python=3.8  # already have current gcc version installed, see below
$ conda activate graphnet
$ pip install -r requirements/torch_cpu.txt -e .[develop,torch]
```
3. See error

**Expected behavior**
GraphNeT is installed.

**Full traceback**
```
Looking in links: https://download.pytorch.org/whl/torch_stable.html, https://data.pyg.org/whl/torch-1.11.0+cpu.html
Obtaining file:///Users/jost/[…]/graphnet
  Preparing metadata (setup.py) ... done
ERROR: Could not find a version that satisfies the requirement torch==1.11+cpu (from versions: 1.4.0, 1.5.0, 1.5.1, 1.6.0, 1.7.0, 1.7.1, 1.8.0, 1.8.1, 1.9.0, 1.9.1, 1.10.0, 1.10.1, 1.10.2, 1.11.0, 1.12.0, 1.12.1, 1.13.0)
ERROR: No matching distribution found for torch==1.11+cpu
```

**Additional context**
`requirements/torch_cpu.txt` requires a version `1.11+cpu`, but the macOS wheels included at https://download.pytorch.org/whl/torch_stable.html are tagged `1.11.0` (without the `+cpu` local version specifier); probably because there are no GPU-specific wheels for macOS? After manually installing the appropriate wheel from the list and commenting out the `torch==1.11+cpu` line in `requirements/torch_cpu.txt`, the installation worked fine otherwise.
