### 2022-12-13 graphnet-team/graphnet#373 @asogaard
Hi @JostMigenda! 👋

Thanks for reviewing, and for filing this bug report. We haven't (yet) targeted macOS builds, but brilliant that it mostly works. I'll try to add a fix asap. (You should also be able to use the Docker image to access IceTray even as a non-IceCube collaborator.)
