### 2022-12-13 graphnet-team/graphnet#373 @JostMigenda
FWIW, I get a 404 error on https://github.com/icecube/icetray/; which I think indicates it’s a private repo? So if the instructions for using IceTray via the Docker image are in that repo, then I can’t access them.

(Also, I did try the “Installing with IceTray” instructions from the GraphNeT README file on our local computing cluster at KCL; but the initial `setup.sh` already fails:
```bash
$ eval `/cvmfs/icecube.opensciencegrid.org/py3-v4.1.0/setup.sh`
/cvmfs/icecube.opensciencegrid.org/py3-v4.1.0/setup.sh: 1: /cvmfs/icecube.opensciencegrid.org/py3-v4.1.0/Ubuntu_20.04_x86_64/bin/python: not found
```
Looks like the CVMFS version doesn’t support Ubuntu 20.04 yet, which our cluster uses. So I’ll try to continue the review with the stand-alone version for now.)
