### 2022-12-13 graphnet-team/graphnet#374 @JostMigenda
Example scripts fail after stand-alone install
As part of https://github.com/openjournals/joss-reviews/issues/4971 I have installed GraphNeT as stand-alone (without IceTray) both on my personal laptop (macOS 12.6.1) and on our university cluster (Ubuntu 20.04) and am trying to run some of the GraphNeT example scripts included in the repo.

Unfortunately, I can use neither of the first four example scripts. On both systems, I get the following errors:
* `compare_sqlite_and_parquet.py`
```bash
$ python examples/compare_sqlite_and_parquet.py
graphnet: INFO     2022-12-13 17:21:29 - get_logger - Writing log to logs/graphnet_20221213-172129.log
graphnet: WARNING  2022-12-13 17:21:31 - warn_once - `icecube` not available. Some functionality may be missing.
Traceback (most recent call last):
  File "examples/compare_sqlite_and_parquet.py", line 78, in <module>
    convert_data()
  File "examples/compare_sqlite_and_parquet.py", line 45, in convert_data
    SQLiteDataConverter(**opt)(paths)  # type: ignore[arg-type]
  File "/users/k2032402/graphnet/graphnet/src/graphnet/data/dataconverter.py", line 188, in __init__
    icetray.I3Logger.global_logger = icetray.I3NullLogger()
NameError: name 'icetray' is not defined
```
* `convert_i3_files.py` fails with the same `NameError` (also in `data/dataconverter.py`, line 188)
* `contour_fitting.py` contains some hard-coded paths like `/mnt/scratch/rasmus_orsoe/…` or `/home/iwsatlas1/oersoe/…` and additionally fails with a `ModuleNotFoundError`:
```bash
$ python examples/contour_fitting.py
Traceback (most recent call last):
  File "examples/contour_fitting.py", line 3, in <module>
    from graphnet.pisa.fitting import ContourFitter
  File "/users/k2032402/graphnet/graphnet/src/graphnet/pisa/fitting.py", line 17, in <module>
    import pisa
ModuleNotFoundError: No module named 'pisa'
```
* `contour_plotting.py` again contains some hard-coded paths like `/mnt/scratch/rasmus_orsoe/…` or `/home/iwsatlas1/oersoe/…`
* (I’ve stopped trying after those four; but a quick glance at the code indicates that at least the hard-coded paths are present in several other example scripts.)

I understand some example scripts may require IceTray-specific functionality or input files that cannot be included in the GraphNeT repo for various reasons. Still, it is quite frustrating to try several example scripts and run into various errors.
At the very least, it would be useful to include an `examples/README.md` file that lists what examples will run with a stand-alone install and which ones require IceTray and/or non-public input files.
