### 2023-01-03 #4971 @JostMigenda
Happy new year, everyone! 👋 Here’s a bit of a fundamental question that I’ve been mulling over the holidays; I would love your thoughts on this and on how to proceed.
Basically, GraphNeT is largely reliant on IceTray, which is proprietary software by the IceCube collaboration; though it can be installed stand-alone. Both modes have advantages and disadvantages, which impact this JOSS review. I’ll discuss my understanding of them here; @asogaard, please correct me if I misunderstood something!

## Stand-Alone Installation

I’ve installed the software as stand-alone thus far. This has worked in principle (kudos to @asogaard for the very quick response to graphnet-team/graphnet#373!), but I have been unable to run any example scripts (see graphnet-team/graphnet#374).
Looking at the code of GraphNeT itself, it appears that this is not a simple bug in the script, but that major parts of functionality are not available in this mode. For example, `graphnet.deployment` currently has a single sub-module `[i3modules](https://github.com/graphnet-team/graphnet/tree/main/src/graphnet/deployment/i3modules)`, which requires imports from `icecube.icetray` to function; so it’s apparently not possible to use any trained GraphNeT models in stand-alone mode.

Stand-alone installation is valuable to ensure that GraphNeT is usable without IceTray and can be adopted by other experiments. This is a good example for how to decouple Open Source software from a proprietary dependency so I commend you for the work on this!
At the same time, adopting GraphNeT for use with another experiment probably requires months of work to integrate with that experiment’s data formats and tools; so it is not feasible to test this as part of a JOSS review.

<summary><strong>Tangent Regarding Adoption by Other Experiments</strong></summary>

Let’s say I work on JANE (Jost’s Amazing Neutrino Experiment) with its TARZAN toolchain (Totally Awesome Reconstruction Zoftware for Astrophysical Neutrinos)—what do I need to do to adopt GraphNeT?

To return to the example above—if I write a `graphnet.deployment` submodule to integrate into TARZAN, does GraphNeT require any specific code structure, class or method names, etc.? Right now, I think I would need to look at IceCube-specific submodules as an example; but if I’m not already familiar with IceTray, it may not be clear which parts of the submodule design are required by GraphNeT and which are IceTray-specific and may not make sense for me.

## Installation with IceTray

IceTray is IceCube-proprietary software. Non-members of IceCube can apparently access it in principle, but [I’ve run into issues while trying](https://github.com/graphnet-team/graphnet/issues/373#issuecomment-1349114191). (@asogaard, I think that should be opened as an issue in the [IceTray repo](https://github.com/icecube/icetray/) (which I cannot do, since it appears to be private); but please let me know if you’d like me to file an issue in the GraphNeT repo about that, even if it’s just so you can refer to it in internal discussions.)

Most parts of the functionality should be available in this mode, so this is suitable for testing as part of this JOSS review. The proprietary dependencies are not ideal, of course, but looking at JOSS’ [policy on proprietary languages and development environments](https://joss.readthedocs.io/en/latest/submitting.html#submissions-using-proprietary-languages-development-environments), that’s not a dealbreaker.

## Next Steps

Since the stand-alone mode is not reasonably testable as part of a JOSS review, I think the best way forward is for me to test the IceTray mode only and have a high-level look at the code itself to ensure there aren’t any major issues in stand-alone mode and that documentation is reasonably complete and accurate. @dfm, does that sound alright?

Independent of how we proceed with this review, I think the overall issue requires a bit more documentation: What GraphNeT functionality (example scripts, submodules, …) is available in which mode? When should I use which mode? (Is it as simple as using IceTray mode to work on IceCube and stand-alone mode to work on another experiment? Or are there cases where I should use IceTray even if I don’t work on IceCube?)
Right now, the README file says that the main difference between the two installation modes is support for I3 files, but that example scripts should work either way; implying that functionality is otherwise largely identical. I was quite surprised initially to find that’s not the case.
