### 2023-01-04 #4971 @asogaard
Hi @dfm, @JostMigenda,

Happy New Years! 😊  I will be on my other gig for the rest of the week, but I will try to provide inputs here in a timely fashion. For now, just a quick note on @JostMigenda's points above:

**Dependence on IceTray**
It is true that parts of GraphNeT depend on IceTray: `graphnet.data` and `graphnet.deployment`,  specifically; whereas `graphnet.models` and `graphnet.training` don't. This is a deliberate choice to facilitate the end-to-end physics pipeline, from experiment-specific data to deployment in experiment-specific code. We're working on onboarding other experiments (P-ONE, KM3NeT), which would mean extending these two modules with the requisite experiment-specific code. But the "main" parts of the package — the parts that actually build and train GNNs — are decoupled from IceTray. Based on @JostMigenda's [request](https://github.com/graphnet-team/graphnet/issues/374), I am [working on](https://github.com/graphnet-team/graphnet/pull/378) restructuring our example scripts to make it more clear which depend on other software and which are entirely stand-alone, as well as provide additional experiment-independent test data for use in these examples.

**Installation**
The stand-alone installation is simply a subset of the installation with IceTray. Installing without IceTray simply means that the IceCube-specific data conversion and deployment is not available (see above). We also have a [Docker image](https://hub.docker.com/repository/docker/asogaard/graphnet) that comes with IceTray and GraphNeT installed together, so if you're keen to try the IceTray build, @JostMigenda, I think that would be the easiest way to do it. (We don't use it that much ourselves currently, so it might be a wee bit rusty, but I'd be happy to assist in getting it working for you.)
