### 2023-01-04 #4971 @dfm
Thanks @JostMigenda for this thorough summary and interesting questions!! And thanks to @asogaard for the response!

It sounds like after https://github.com/graphnet-team/graphnet/pull/378 has been merged, it would be possible to properly review the standalone version as well. My (weak!) preference would be to try to provide feedback on both the standalone and the version including the public binary of IceTray if possible, but if that's not feasible or timely, I'd be perfectly happy with just a review for the full version, as long as the docs are clear about any limitations of the standalone version. This is exactly the kind of issue that we're looking to diagnose with these JOSS reviews so thanks both for all your work on this!!
