### 2023-01-04 #4971 @GageDeZoort
Apologies for my delayed response here - my feedback largely echoes @JostMigenda's comments above, specifically that the examples need more thorough documentation about what's required to run them. So far, I've had no trouble installing the standalone GraphNeT, but had similar trouble running the examples due to their dependencies on IceTray.

The changes outlined so far (e.g. [graphnet-team/graphnet#378](https://github.com/graphnet-team/graphnet/pull/378)) make a lot of sense. I'd also suggest adding (at the top of each example .py file or in a bulleted list in the ReadMe) info about how to run each script, what the expected output is, hard-coded lines that need to be changed (e.g. W&B config info), and what dependencies are required to run the script (stand-alone vs. IceTray).

I have not yet tried the IceTray-dependent installation, but would be happy to try using the IceTray+GraphNeT Docker image as @dfm and @asogaard suggestion. I'll look more into this in the afternoon.
