### 2023-01-20 #4971 @asogaard
Hi @dfm, @JostMigenda, and @GageDeZoort,

A brief status update: We just merged https://github.com/graphnet-team/graphnet/pull/378. It was a considerable effort, but I think the package is all the better for it. Thanks for the suggestions, @JostMigenda and @GageDeZoort!

* All examples are now **structured into sub-folders** representing different steps or purposes in a normal workflow. This also serves to group examples with similar dependencies (IceTray or PISA).
* All examples are now equipped with a **simple command-line interface**, that explains how to run the script.
* All examples for which the user has the required packages installed should now **run out of the box**, using small example datasets that come bundled with the package.
* All examples for which the user **doesn't** have the required packages installed now produce **an informative error message.**
* I have added a few **README's,** specifically in the top-level examples directory, in the IceTray- and PISA-specific examples directories (explaining the dependency on a non-standard python package), and in the training-specific examples directory, as this is by far the most intersting one.

Please let me know if this does the job, once you get a chance to have a second look. I will move on to have a look at the Docker image to make sure it is still works as intended. I will report back here once ready.
