### 2023-01-26 graphnet-team/graphnet#392 @JostMigenda
Logging of multiple workers behaves unpredictably
**Describe the bug**
Some tasks (e.g. loading data) spawn a large number of worker processes to parallelize the workload. Every worker process appears to determine their log file name based on the current datetime with second-level resolution, so multiple worker processes started at (almost) the same time may write to the same log file or different log files unpredictably.

**To Reproduce**
Steps to reproduce the behavior:
1. Start clean shell.
2. Activate environment with stand-alone install of GraphNeT. (I have not tested with an IceTray install yet.)
3. Run `python examples/02_data/01_read_dataset.py sqlite`
4. Run `ls logs/` and look through log files created by this script run.

For example, in one case, the main script wrote logs to `logs/graphnet_20230126-162251.log`; of the 30 workers, 4 wrote their logs to `logs/graphnet_20230126-162300.log`, 20 wrote to `logs/graphnet_20230126-162301.log` and the remaining 6 wrote to `logs/graphnet_20230126-162302.log`.
The exact details vary; when using `parquet` instead of `sqlite`, worker startup time seems longer and I’ve seen the worker logs split into as many as 7 log files.

**Expected behavior**
I see two options here:

1. All workers use the same log file as the parent process that started them. If this option is selected, every log statement should indicate which worker it was produced by; otherwise log messages from multiple workers may be interspersed, making it hard to understand the logs.
2. Every worker uses an individual log file. (Though this may create an impractically large number of log files.)

**Full traceback**
N/A

**Additional context**
Discovered as part of openjournals/joss-reviews#4971. (Though to be clear: I’m aware that herding parallel worker process can become surprisingly hairy; so if this is one of those cases, we shouldn’t delay the JOSS review for what’s a minor usability issue in most cases.)
