### 2023-01-26 graphnet-team/graphnet#393 @JostMigenda
Example script `02_data/02_plot_feature_distributions.py` fails
**Describe the bug**
As part of openjournals/joss-reviews#4971, I’ve started going through the example script. On my machine, running `02_data/02_plot_feature_distributions.py` fails; probably because `configs/datasets/test_data_sqlite.yml` points to an inaccessible file (`/groups/icecube/asogaard/…`).

**To Reproduce**
Steps to reproduce the behavior:
1. Start clean shell (on a machine that doesn’t have access to `/groups/icecube/asogaard/…`)
2. Activate environment with stand-alone install of GraphNeT. (I have not tested with an IceTray install yet.)
3. Run `python examples/02_data/02_plot_feature_distributions.py`
4. See error

**Expected behavior**
Script runs successfully and produces output.

**Full traceback**
Please include the full error message to allow for debugging
```
graphnet: INFO     2023-01-26 17:12:10 - get_logger - Writing log to logs/graphnet_20230126-171210.log
graphnet: WARNING  2023-01-26 17:12:12 - warn_once - `icecube` not available. Some functionality may be missing.
Traceback (most recent call last):
  File "examples/02_data/02_plot_feature_distributions.py", line 94, in <module>
    main()
  File "examples/02_data/02_plot_feature_distributions.py", line 20, in main
    dataset = Dataset.from_config("configs/datasets/test_data_sqlite.yml")
  File "/…/graphnet/src/graphnet/data/dataset.py", line 60, in from_config
    return source._dataset_class(**source.dict())
  File "/…/graphnet/src/graphnet/utilities/config/dataset_config.py", line 164, in wrapper
    ret = init_fn(self, *args, **kwargs)
  File "/…/graphnet/src/graphnet/data/dataset.py", line 244, in __init__
    self._indices = self._get_all_indices()
  File "/…/graphnet/src/graphnet/data/sqlite/sqlite_dataset.py", line 83, in _get_all_indices
    self._establish_connection(0)
  File "/…/graphnet/src/graphnet/data/sqlite/sqlite_dataset.py", line 111, in _establish_connection
    self._conn = sqlite3.connect(self._path)
sqlite3.OperationalError: unable to open database file
```

**Additional context**
There are several other hardcoded paths to non-public files like that. I haven’t run into issues with the others yet; but perhaps this is a good opportunity to double-check if any of those should be fixed, too.

```
> igrep groups/icecube configs/datasets/*
configs/datasets/dev_lvl7_robustness_muon_neutrino_0000.yml:path: /groups/icecube/asogaard/data/example/dev_lvl7_robustness_muon_neutrino_0000.db
configs/datasets/PID_classification_last_one_lvl3MC.yml:path: /groups/icecube/petersen/GraphNetDatabaseRepository/Leon2022_DataAndMC_CSVandDB_StoppedMuons/last_one_lvl3MC.db
configs/datasets/test_data_sqlite.yml:path: /groups/icecube/asogaard/work/development/graphnet/data/tests/sqlite/oscNext_genie_level7_v02/oscNext_genie_level7_v02_first_5_frames.db
configs/datasets/training_example_data_parquet.yml:path: /groups/icecube/asogaard/work/development/graphnet/data/examples/parquet/prometheus/prometheus-events.parquet
configs/datasets/training_example_data_sqlite.yml:path: /groups/icecube/asogaard/work/development/graphnet/data/examples/sqlite/prometheus/prometheus-events.db
```

----

Additionally, when running `cd examples/` before step 3, there is a separate error message:
```
Traceback (most recent call last):
  File "02_data/02_plot_feature_distributions.py", line 94, in <module>
    main()
  File "02_data/02_plot_feature_distributions.py", line 20, in main
    dataset = Dataset.from_config("configs/datasets/test_data_sqlite.yml")
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataset.py", line 43, in from_config
    source = DatasetConfig.load(source)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/utilities/config/base_config.py", line 25, in load
    with open(path, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: 'configs/datasets/test_data_sqlite.yml'
```
Personally, I would find it nice to be able to run all example scripts from within the `examples/` directory (or even one of the numbered subdirectories) so the `logs/` and `outputs/` directories don’t pollute the root directory of the repo; but if they need to be run from a specific directory, that should be mentioned explicitly in the README or that script’s `--help` text.
