### 2023-01-26 graphnet-team/graphnet#393 @JostMigenda
Just noticed that the same issue occurs in `examples/04_training/01_train_model.py` when running it with the default dataset config (`--dataset-config configs/datasets/training_example_data_sqlite.yml`):
```
Traceback (most recent call last):
  File "examples/04_training/01_train_model.py", line 159, in <module>
    main(
  File "examples/04_training/01_train_model.py", line 70, in main
    dataloaders = DataLoader.from_dataset_config(
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataloader.py", line 66, in from_dataset_config
    datasets = Dataset.from_config(config)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataset.py", line 52, in from_config
    return cls._construct_datasets_from_dict(source)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataset.py", line 80, in _construct_datasets_from_dict
    dataset = Dataset.from_config(config)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataset.py", line 60, in from_config
    return source._dataset_class(**source.dict())
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/utilities/config/dataset_config.py", line 164, in wrapper
    ret = init_fn(self, *args, **kwargs)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataset.py", line 246, in __init__
    self._indices = self._resolve_string_selection_to_indices(
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/dataset.py", line 356, in _resolve_string_selection_to_indices
    return self._string_selection_resolver.resolve(selection)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/utilities/string_selection_resolver.py", line 95, in resolve
    df_selection = self._query_selection_from_dataset(selection)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/utilities/string_selection_resolver.py", line 215, in _query_selection_from_dataset
    data=self._dataset.query_table(
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/sqlite/sqlite_dataset.py", line 61, in query_table
    self._establish_connection(index)
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/data/sqlite/sqlite_dataset.py", line 111, in _establish_connection
    self._conn = sqlite3.connect(self._path)
sqlite3.OperationalError: unable to open database file
```

I also noticed that that script runs from any folder by using [`graphnet.constants.CONFIG_DIR`](https://github.com/graphnet-team/graphnet/blob/dbaa88506b2d61f91ff3040792bf89f8b8c69ee3/src/graphnet/constants.py#L33) instead of hard-coding the relative path to the config file.
