### 2023-01-26 graphnet-team/graphnet#394 @JostMigenda
Example script `04_training/02_train_model_without_configs.py` fails
**Describe the bug**
As part of openjournals/joss-reviews#4971, I’ve started going through the example script. On my machine, running `04_training/02_train_model_without_configs.py` fails with `AttributeError: Can't pickle local object 'make_dataloader.<locals>.collate_fn'` somewhere deep in the innards of the `multiprocessing` module.

**To Reproduce**
Steps to reproduce the behavior:
1. Start clean shell.
2. Activate environment with stand-alone install of GraphNeT. (I have not tested with an IceTray install yet.)
3. Run `python examples/04_training/02_train_model_without_configs.py`
4. See error

**Expected behavior**
Script runs successfully and produces output.

**Full traceback**
<summary>Expand this to view full traceback (it’s a long one …)</summary>

```
> python examples/04_training/02_train_model_without_configs.py
graphnet: INFO     2023-01-26 18:13:54 - get_logger - Writing log to logs/graphnet_20230126-181354.log
graphnet: WARNING  2023-01-26 18:13:55 - warn_once - `icecube` not available. Some functionality may be missing.
wandb: (1) Create a W&B account
wandb: (2) Use an existing W&B account
wandb: (3) Don't visualize my results
wandb: Enter your choice: 3
wandb: You chose 'Don't visualize my results'
wandb: WARNING `resume` will be ignored since W&B syncing is set to `offline`. Starting a new run with run id 18kje1tx.
wandb: Tracking run with wandb version 0.13.6
wandb: W&B syncing is set to `offline` in this directory.
wandb: Run `wandb online` or set WANDB_MODE=online to enable cloud syncing.
graphnet: INFO     2023-01-26 18:14:01 - main - features: ['sensor_pos_x', 'sensor_pos_y', 'sensor_pos_z', 't']
graphnet: INFO     2023-01-26 18:14:01 - main - truth: ['injection_energy', 'injection_type', 'injection_interaction_type', 'injection_zenith', 'injection_azimuth', 'injection_bjorkenx', 'injection_bjorkeny', 'injection_position_x', 'injection_position_y', 'injection_position_z', 'injection_column_depth', 'primary_lepton_1_type', 'primary_hadron_1_type', 'primary_lepton_1_position_x', 'primary_lepton_1_position_y', 'primary_lepton_1_position_z', 'primary_hadron_1_position_x', 'primary_hadron_1_position_y', 'primary_hadron_1_position_z', 'primary_lepton_1_direction_theta', 'primary_lepton_1_direction_phi', 'primary_hadron_1_direction_theta', 'primary_hadron_1_direction_phi', 'primary_lepton_1_energy', 'primary_hadron_1_energy', 'total_energy']
GPU available: False, used: False
TPU available: False, using: 0 TPU cores
IPU available: False, using: 0 IPUs
HPU available: False, using: 0 HPUs
Trainer already configured with model summary callbacks: [<class 'pytorch_lightning.callbacks.model_summary.ModelSummary'>]. Skipping setting a default `ModelSummary` callback.
GPU available: False, used: False
TPU available: False, using: 0 TPU cores
IPU available: False, using: 0 IPUs
HPU available: False, using: 0 HPUs
Initializing distributed: GLOBAL_RANK: 0, MEMBER: 1/1
----------------------------------------------------------------------------------------------------
distributed_backend=gloo
All distributed processes registered. Starting with 1 processes
----------------------------------------------------------------------------------------------------

  | Name      | Type       | Params
-----------------------------------------
0 | _detector | Prometheus | 0
1 | _gnn      | DynEdge    | 1.4 M
2 | _tasks    | ModuleList | 129
-----------------------------------------
1.4 M     Trainable params
0         Non-trainable params
1.4 M     Total params
5.515     Total estimated model params size (MB)
Sanity Checking: 0it [00:00, ?it/s]Traceback (most recent call last):
  File "examples/04_training/02_train_model_without_configs.py", line 210, in <module>
    main(
  File "examples/04_training/02_train_model_without_configs.py", line 135, in main
    model.fit(
  File "/Users/jost/Documents/Academia/Papers/Reviews/JOSS GraphNeT/graphnet/src/graphnet/models/model.py", line 80, in fit
    trainer.fit(
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 603, in fit
    call._call_and_handle_interrupt(
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/call.py", line 36, in _call_and_handle_interrupt
    return trainer.strategy.launcher.launch(trainer_fn, *args, trainer=trainer, **kwargs)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/strategies/launchers/subprocess_script.py", line 90, in launch
    return function(*args, **kwargs)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 645, in _fit_impl
    self._run(model, ckpt_path=self.ckpt_path)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 1098, in _run
    results = self._run_stage()
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 1177, in _run_stage
    self._run_train()
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 1190, in _run_train
    self._run_sanity_check()
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 1262, in _run_sanity_check
    val_loop.run()
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/loops/loop.py", line 199, in run
    self.advance(*args, **kwargs)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/loops/dataloader/evaluation_loop.py", line 152, in advance
    dl_outputs = self.epoch_loop.run(self._data_fetcher, dl_max_batches, kwargs)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/loops/loop.py", line 194, in run
    self.on_run_start(*args, **kwargs)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/loops/epoch/evaluation_epoch_loop.py", line 84, in on_run_start
    self._data_fetcher = iter(data_fetcher)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/pytorch_lightning/utilities/fetching.py", line 178, in __iter__
    self.dataloader_iter = iter(self.dataloader)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/torch/utils/data/dataloader.py", line 363, in __iter__
    self._iterator = self._get_iterator()
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/torch/utils/data/dataloader.py", line 314, in _get_iterator
    return _MultiProcessingDataLoaderIter(self)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/site-packages/torch/utils/data/dataloader.py", line 927, in __init__
    w.start()
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/process.py", line 121, in start
    self._popen = self._Popen(self)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/context.py", line 224, in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/context.py", line 284, in _Popen
    return Popen(process_obj)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/popen_spawn_posix.py", line 32, in __init__
    super().__init__(process_obj)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/popen_fork.py", line 19, in __init__
    self._launch(process_obj)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/popen_spawn_posix.py", line 47, in _launch
    reduction.dump(process_obj, fp)
  File "/Users/jost/opt/anaconda3/envs/graphnet/lib/python3.8/multiprocessing/reduction.py", line 60, in dump
    ForkingPickler(file, protocol).dump(obj)
AttributeError: Can't pickle local object 'make_dataloader.<locals>.collate_fn'
```

**Additional context**
I use GraphNeT 0.2.4+25.gdbaa885 and Python 3.8.15 (installed via conda) on macOS 12.6.3.
Let me know if you need any additional details from me to debug this!
