### 2023-01-26 8cd44e4
Don't hardcode config dir in example/02/02 Allows users to run that example script from different directories, as mentioned as an aside in #393
