### 2023-01-26 #4971 @JostMigenda
Thanks @asogaard for this major reworking of the examples! It indeed looks a lot better now; I’ve gone through most of the example scripts this afternoon and have opened issues for the few cases where I ran into problems. (And it looks like you already fixed one of them before I was done—very speedy! 🚀)

I see the Docker image is ready now; I’ll try to run that in a bit and try out the scripts in `examples/01_icetray/`.
Just one more thing for now: The README file says that PISA isn’t available in the Docker image—is there any particular reason not to include that?
