### 2023-01-26 graphnet-team/graphnet#397 @JostMigenda
Docker issues
As part of openjournals/joss-reviews#4971, I’ve tried to run the Docker image and have a couple of initial notes/issues. (This is just a quick list based on initial impressions; I didn’t want to open lots of small issues immediately, but please feel free to move individual points into their own issue if you think it’s appropriate.)

- [x] The total download size was ~10 GB, which was larger than I expected. I feel like that’s large enough that a warning in the README file might be appropriate.
- [x] Additionally, the README file recommends `docker run --rm […]`, which according to documentation automatically removes the Docker container after it exits. Does that mean I would have to re-download the whole ~10 GB every single time? (I vaguely remember that Docker lets you update layers of a container without having to rebuild lower layers—so e.g. updating from graphnet v0.2.3 to v0.2.4 would only require downloading and rebuilding the graphnet part, but not other parts of the container (Linux, Python, PyTorch, …); that would probably be very useful here.)
- [x] The container downloaded by `asogaard/graphnet:latest` is v0.2.3; [there is no container for v0.2.4](https://hub.docker.com/r/asogaard/graphnet/tags) (which was tagged 11 hours ago) and the GitHub Action that automatically builds & uploads containers [appears to be broken](https://github.com/graphnet-team/graphnet/actions/runs/4014734888) currently.
- [x] Once the container is built, instead of `~/graphnet` (as in the README) I find myself in `~` and there is no `graphnet` directory. (GraphNeT itself is installed in `/usr/local/lib/python3.8/dist-packages/graphnet/`; but of course that doesn’t include the examples directory.) Instead, there are only:
```bash
root@c77c9a429624:~# ls
cppzmq-master/  master.tar.gz  photospline-2.0.7/
```
