### 2023-01-27 graphnet-team/graphnet#397 @asogaard
Aye, this is known, cf. [here](https://github.com/graphnet-team/graphnet/actions/workflows/docker.yml) where the build producing the image failed from v0.2.4 onwards. But working on it over [here](https://github.com/asogaard/graphnet/tree/update-docker), addressing all four issues above, and will create a PR soon. :)
