### 2023-01-28 graphnet-team/graphnet#397 @asogaard
Alright, following the merge of #399 we have a new `main` docker image [here](https://hub.docker.com/layers/asogaard/graphnet/main/images/sha256-cf49a193a0aaee5eb2636e483dfad1c391fe735291d4132e164a96349b35d0d1?context=repo):

* It's 1.59 GB in size cf. approx. 10 GB before. This is down to the base [icetray image](https://hub.docker.com/r/icecube/icetray) used, where we have switched from `combo-stable` to `combo-stable-slim`.
* Using the `--rm` option with `docker run` doesn't remove the downloaded _image_, just the running _container_ and it's filesystem. (The idea is that if you do `docker run` many times without `--rm?  , you create lots of persistent containers that quickly clutter a lot). So you shouldn't have to redownload the _X_ GBs with every run. But I can try to make clear in the instructions that this is entirely optional — and that in some cases you might want to keep the container after exiting.
* Yeah, `v0.2.4` broke. The `asogaard/graphnet:main` container now works as intended. Once you all have had a chance to check that it looks fine, I'll create another tag (`v0.2.5`) such that the `latest` tag will also be the updated version.
* The file system should look as indicated in the README.

So make sure that you pull the `main` version from Dockerhub, and let me know if you consider the four tasks above resolved. (I'll leave it to you to check them off.)
