### 2023-01-30 graphnet-team/graphnet#400 @RasmusOrsoe
Polish Documentation landing Page
The landing page is very sparse. A bit of graphics and a hello would be nice.

Webpage: https://graphnet-team.github.io/graphnet/index.html

**Additional context**
We could seek inspiration from https://pytorch.org/docs/stable/index.html or other larger libraries.
