### 2023-01-30 graphnet-team/graphnet#401 @RasmusOrsoe
Write developer notes for improved documentation of key parts of the library
Larger libraries tend to have "developer notes". They tend to include explanation of key parts of the library in a article-sized format.  I think it might be worthwhile for us to consider if we should do the same for 1.0.0.

**Additional context**
See https://pytorch.org/docs/stable/index.html for examples on developer notes.
