### 2023-01-30 graphnet-team/graphnet#397 @JostMigenda
I’ve just tested `asogaard/graphnet:main` and it looks good now. Thanks also for clearing up my misunderstanding of `--rm`!
