### 2023-01-30 graphnet-team/graphnet#402 @JostMigenda
Crash in `01_icetray/01_convert_i3_files.py` example script
**Describe the bug**
As part of openjournals/joss-reviews#4971, I’m trying out the example scripts in the Docker image and have discovered a crash in the example script `01_icetray/01_convert_i3_files.py` when running with the arguments `parquet icecube-upgrade`.

Additionally, the output directory `{EXAMPLE_OUTPUT_DIR}/convert_i3_files/ic86/` is used for both values of `detector` (`icecube-86` and `icecube-upgrade`), which seems like it might not be intended?

**To Reproduce**
Steps to reproduce the behavior:
1. Run Docker image as `docker run --rm -it asogaard/graphnet:main`
2. In the container, run example script: `python examples/01_icetray/01_convert_i3_files.py parquet icecube-upgrade`
3. See error

**Expected behavior**
I expect the script to run successfully and produce the expected output files, as it does when running `python examples/01_icetray/01_convert_i3_files.py parquet icecube-86`.

**Full traceback**
```bash
🐳 graphnet@fccc2bd071dd ❯ ~/graphnet $ python examples/01_icetray/01_convert_i3_files.py parquet icecube-upgrade
graphnet: INFO     2023-01-30 13:04:00 - get_logger - Writing log to logs/graphnet_20230130-130400.log
/usr/local/lib/python3.8/dist-packages/pandas/core/computation/expressions.py:20: UserWarning: Pandas requires version '2.7.3' or newer of 'numexpr' (version '2.7.1' currently installed).
  from pandas.core.computation.check import NUMEXPR_INSTALLED
graphnet: INFO     2023-01-30 13:04:03 - ParquetDataConverter.info - Saving results to /root/graphnet/data/examples/outputs/convert_i3_files/ic86
graphnet: INFO     2023-01-30 13:04:03 - ParquetDataConverter.info - Processing 1 I3 file(s) in main thread (not multiprocessing)
  0%|                                                                                                                                                                                                                         | 0/1 [00:12<?, ?file(s)/s]
Traceback (most recent call last):
  File "examples/01_icetray/01_convert_i3_files.py", line 119, in <module>
    main_icecube_upgrade(args.backend)
  File "examples/01_icetray/01_convert_i3_files.py", line 80, in main_icecube_upgrade
    converter(inputs)
  File "/root/graphnet/src/graphnet/data/dataconverter.py", line 215, in __call__
    self.execute(filesets)
  File "/root/graphnet/src/graphnet/data/dataconverter.py", line 286, in execute
    pool = self._iterate_over_individual_files(filesets)
  File "/root/graphnet/src/graphnet/data/dataconverter.py", line 330, in _iterate_over_individual_files
    for _ in map_fn(
  File "/root/graphnet/src/graphnet/data/dataconverter.py", line 83, in wrapper
    output_file = process_method(self, *args)
  File "/root/graphnet/src/graphnet/data/dataconverter.py", line 381, in _process_file
    self.save_data(data, output_file)
  File "/root/graphnet/src/graphnet/data/parquet/parquet_dataconverter.py", line 32, in save_data
    awkward.to_parquet(awkward.from_iter(data), output_file)
  File "/usr/local/lib/python3.8/dist-packages/awkward/operations/convert.py", line 3071, in to_parquet
    writer = pyarrow.parquet.ParquetWriter(**options)
  File "/usr/local/lib/python3.8/dist-packages/pyarrow/parquet/core.py", line 990, in __init__
    self.writer = _parquet.ParquetWriter(
  File "pyarrow/_parquet.pyx", line 1753, in pyarrow._parquet.ParquetWriter.__cinit__
  File "pyarrow/error.pxi", line 144, in pyarrow.lib.pyarrow_internal_check_status
  File "pyarrow/error.pxi", line 100, in pyarrow.lib.check_status
pyarrow.lib.ArrowInvalid: NullType Arrow field must be nullable
```

**Additional context**
n/a
