### 2023-01-30 #4971 @JostMigenda
I tested the Docker image in the last few days and ran the IceTray examples; apart from one remaining bug, this looks quite good now. Once the fixes in graphnet-team/graphnet#396 are merged, I think the paper is also ready.
I think we’re getting fairly close to the end of my review; here’s a list of all open questions that I see remaining right now:

- [ ] Testing of the PISA scripts: It would be nice if PISA were included in the Docker image; but if there’s any particular reason not to include it and I should try to install PISA manually instead, let me know.
- [X] The `graphnet.deployment` submodule is not exercised by any of the examples. As discussed above, this is experiment-specific, so I’d be okay with that; just wanted to put this on the record.
- [ ] I think the [online documentation](https://graphnet-team.github.io/graphnet/index.html) needs a bit more work. Sphinx’ autogenerated index pages are useful as an API reference for experienced GraphNeT developers; but for users such as myself, who do not have much experience using GraphNeT (yet), they’re overwhelming and lacking structure. The example scripts are very nice as a basic introduction, but I think there is still a gap that needs to be filled.
Standalone articles (as proposed in graphnet-team/graphnet#401) would be great; but I realize writing those well is a lot of work. If you don’t think that’s achievable on the time scale of this review, here’s my suggestion for an interim solution that wouldn’t require much additional work:
    * On the landing page (as already suggested in graphnet-team/graphnet#400), add a pointer to the README file (for installation instructions etc.) and a brief explanation of what GraphNeT is—in particular, you could include the figure from the JOSS paper which gives a nice and clear overview over the structure of GraphNeT.
    * Add a separate page for each submodule to the TOC, so the structure of the documentation (which is clearly visible in the sidebar) reflects the structure of GraphNeT and users don’t have to comb through the alphabetical index of all sub-(sub-)modules to find what they’re looking for.
    * On the page for each submodule, structure and clean up the autogenerated content a bit.[*]
    * Finally, go through the autogenerated pages and see where additional explanations would be useful in the module-level docstrings. (For example, is there an existing example script that illustrates the usage of this submodule nicely? Is it clear which classes are intended for general use and which ones are IceCube-proprietary? Etc.)

[*] A few examples of what I would do to improve [the autogenerated `graphnet.training` page](https://graphnet-team.github.io/graphnet/graphnet.training.html):
* The module-level docstring should be at the top of the page, not at the bottom.
* Most users who read the `graphnet.training` page probably aren’t interested in a [helper class to modify the appearance of progress bars](https://graphnet-team.github.io/graphnet/graphnet.training.html#graphnet.training.callbacks.ProgressBar), so it would be better to hide that.
* For submodules which have a base class and a bunch of derived classes (e.g. `graphnet.training.loss_functions`, `graphnet.training.weight_fitting`), it would be clearer to have the base class (which documents the common interface) at the top and include all derived classes below that.
* Private functions that contain implementation details and are not intended for users (typically starting with an underscore) should usually not be included in user-facing docs.

In my experience, `sphinx.ext.autodoc`’s `automodule` and `autoclass` directives are flexible enough that it should be possible to make all those changes while still keeping most of the benefits of completely auto-generated docs.
