### 2023-01-31 graphnet-team/graphnet#404 @asogaard
Make PISA examples self-contained
With #378 most examples scripts are self-contained, i.e. referring only to files that come bundled with the repo. The only exception is the PISA ones.

**Describe the solution you'd like**
It would be nice if all the example scripts we provide, including the PISA ones, could be run out of the box, provided of course that the necessary packages are installed. This is also requested here: https://github.com/openjournals/joss-reviews/issues/4971#issuecomment-1408985970.

If we want to keep our examples self-contained, the alternative would be to not provide examples on the interface with PISA, but I think this would be a shame.
