### 2023-01-31 graphnet-team/graphnet#405 @asogaard
Install PISA in main Docker image
Right now the main `graphnet` docker image comes with `graphnet` and `icetray` installed, which allows for running most example scripts out of the box. However, currently PISA is not installed, which means that the PISA examples scripts (even if they were self-contained; cf. #404) cannot be run out of the box in the provided Docker image, which can be frustrating. See also https://github.com/openjournals/joss-reviews/issues/4971#issuecomment-1408985970.

**Describe the solution you'd like**
It would be nice to provide a Docker image that comes with all the most relevant packages installed, including ancillary ones like PISA.

Alternatively, we should decide that it is not within the scope of `graphnet` proper to support an interface with PISA (arguing, e.g., that it is a very specialised end-use case) and move this functionality out of the main repository. Tagging @RasmusOrsoe for a second opinion.
