### 2023-01-31 graphnet-team/graphnet#394 @asogaard
Have tried reproducing the above error after the merging of #408, and the crash seems to have been solved. Closing, but feel free to reopen, @JostMigenda, if you find the problem persists.
