### 2023-01-31 graphnet-team/graphnet#405 @JostMigenda
FWIW, if you decide that PISA support is a specialized, IceCube-proprietary feature and not intended for general usage, then I think it's sufficient if you explicitly state that in the module docstring and the examples/04_pisa README.
Please don't feel like you have to remove the code and put it e.g. in a separate repo which might make it harder to use.
