### 2023-02-14 graphnet-team/graphnet#404 @RasmusOrsoe
@asogaard as discussed with you and @philippeller , we cannot do this unless we have significantly higher statistics in our public data. (Running a minimizer on ~5 events wont work). So I think we can close this.
