### 2023-02-14 graphnet-team/graphnet#404 @asogaard
Closing in favour of #428
