### 2023-02-14 #4971 @GageDeZoort
I've been going through the code and compiling some feedback. All-in-all things appear to be running well. I'll post some more detailed items soon, but in the meantime I've got a few high-level questions/comments for the authors:

Overall comments/questions:
- The EdgeConv-based architecture is clearly quite performative, but if this package is billed as a more all-encompassing package, I'd expect to see some additional commonly used GNN layers available. For example, [Interaction Networks](https://github.com/gnn-tracking/gnn_tracking/blob/main/src/gnn_tracking/models/interaction_network.py) pop up frequently at the LHC. There are also FPGA-accelerated variants of EdgeConv (called [GravNet/GarNet](https://arxiv.org/abs/1902.07987)) available, so it could be nice to provide versions of these algorithms here. I think the PyG strategy is best, to provide the layers themselves but to let the users explore how to integrate them into architectures.
- The documentation/paper could benefit from additional clarity about when to interface with external models, clustering algorithms, etc., vs. using the versions provided in GraphNeT.
- Are there plans to include hyperparameter scanner utilities? I'm not convinced they need to be provided here, I'm mainly curious how you thought about this. Perhaps this could also go into the workflow diagram in the paper?
- Can you please explain the choice to provide custom implementations of GNNs instead of leveraging the PyG `MessagePassing` class?
- The documentation is still relatively sparse - it might be nice to populate it with write-ups of some of the example scripts. Otherwise, a write-up (e.g. text surrounding Jupyter notebook cells) showing an example workflow you might hand to a student looking to develop their own GNN might be useful.
