### 2023-03-21 #4971 @asogaard
Hi @dfm, @GageDeZoort, @JostMigenda,

I am just back from parental leave and wanted to circle back to this. It seems like the review checklists are very close to complete. Therefore, and considering the review has been ongoing since November 2022, I'd be keen to push towards acceptance in JOSS while this effort is still a priority to the GraphNeT community.

There has been a number of questions, requests, etc. surfaced by the reviewers over the past few months. In order to allow us to assess the scope of the remaining work and to work effectively towards acceptance, I was wondering whether I could ask @GageDeZoort and @JostMigenda to each **summarise what you see as outstanding issues that must be resolved before they can consider their review complete?**
