### 2023-03-24 #4971 @asogaard
Thanks, @GageDeZoort, I'm providing some answers below. I hope this provides enough context to be able to itemise the remaining acceptance-blockers as you see them.

-----

> * The EdgeConv-based architecture is clearly quite performative, but if this package is billed as a more all-encompassing package, I'd expect to see some additional commonly used GNN layers available. For example, [Interaction Networks](https://github.com/gnn-tracking/gnn_tracking/blob/main/src/gnn_tracking/models/interaction_network.py) pop up frequently at the LHC. There are also FPGA-accelerated variants of EdgeConv (called [GravNet/GarNet](https://arxiv.org/abs/1902.07987)) available, so it could be nice to provide versions of these algorithms here. I think the PyG strategy is best, to provide the layers themselves but to let the users explore how to integrate them into architectures.

The goal is not provide an _all-encompassing_ package, but rather provide a framework which supports the end-to-end use of GNNs and is _easy to extend_ with new layers and architectures. We provide [two](https://github.com/graphnet-team/graphnet/blob/main/src/graphnet/models/gnn/convnet.py) [example](https://github.com/graphnet-team/graphnet/blob/main/src/graphnet/models/gnn/dynedge.py) GNN layers/architectures, and users are more than welcome to contribute new ones, like the two you suggest.

> * The documentation/paper could benefit from additional clarity about when to interface with external models, clustering algorithms, etc., vs. using the versions provided in GraphNeT.

Which external models, clustering algorithms, etc. are you thinking of?

> * Are there plans to include hyperparameter scanner utilities? I'm not convinced they need to be provided here, I'm mainly curious how you thought about this. Perhaps this could also go into the workflow diagram in the paper?

We have users who have used W&B to do hyperparameter sweeps, but while we plan to provide examples of how to do this at some point, it is not considered core functionality.

> * Can you please explain the choice to provide custom implementations of GNNs instead of leveraging the PyG `MessagePassing` class?

I think there is a good argument to be made that our `GNN`-type modules could benefit from leveraging `MessagePassing`. I think this choice was mainly because we primarily wanted consistency across out different module types (`Detector`, `GNN`, `Coarsening`, `Task`, and so on) and not all these modules are message-passing-type and wouldn't easily fit the `MessagePassing.forward` syntax. But for a later release we are thinking about how to make our models `jittable` to support simpler deployment, but that is a ways further down the road.

> * The documentation is still relatively sparse - it might be nice to populate it with write-ups of some of the example scripts. Otherwise, a write-up (e.g. text surrounding Jupyter notebook cells) showing an example workflow you might hand to a student looking to develop their own GNN might be useful.

We are working on a tutorial notebook of sorts that works through the main training example scripts and tries to do a more detailed job of explaining the different bits, how they fit together, and how they can be changed.
