### 2023-03-28 #4971 @asogaard
Hi @dfm, @GageDeZoort, @JostMigenda,

I just wanted to follow up on this to check whether, in addition to the answers to @GageDeZoort's questions above, there is anything you need from me in order to itemise the remaining acceptance-blockers as you see them?
