### 2023-03-28 #4971 @JostMigenda
Sorry for the slow response; between strikes at UK universities and an upcoming conference I’m co-organising, I’ve got a bit of a backlog right now. Congratulations on the new kid, @asogaard! Hope you enjoyed the parental leave and are occasionally getting some sleep … 😉

The only significant issue I still see is related to the dependence on IceTray. For `graphnet.data`, I think the experiment-independent examples are very helpful, so that’s resolved. However, for `graphnet.deployment` I’m not sure the current status is clear enough. To quote my tangent from [an earlier post](https://github.com/openjournals/joss-reviews/issues/4971#issuecomment-1370124479):

> Let’s say I work on JANE (Jost’s Amazing Neutrino Experiment) with its TARZAN toolchain (Totally Awesome Reconstruction Zoftware for Astrophysical Neutrinos)—what do I need to do to adopt GraphNeT?
>
> To return to the example above—if I write a `graphnet.deployment` submodule to integrate into TARZAN, does GraphNeT require any specific code structure, class or method names, etc.? Right now, I think I would need to look at IceCube-specific submodules as an example; but if I’m not already familiar with IceTray, it may not be clear which parts of the submodule design are required by GraphNeT and which are IceTray-specific and may not make sense for me.

Ideally, there would be full documentation on this that allows a researcher on a new experiment to deploy GraphNeT without requiring help from GraphNeT developers. In practice, due to the complexity of operating such a neutrino telescope, maybe it’s reasonable to say that integration details can’t be documented well because they differ too much between experiments? (You said that you’re already working with P-ONE and KM3NeT—maybe you have some experience of how much this differs already?)

If you say that, in practice, it will almost certainly be necessary for a new experiment to collaborate directly with GraphNeT developers anyway, then at the very least I think it would be good to have a brief explanation in the [`graphnet.deployment` docstring](https://github.com/graphnet-team/graphnet/blob/main/src/graphnet/deployment/__init__.py). (Very roughly: “Deployment submodules are closely linked to each experiment’s proprietary software/DAQ/…, so they will differ significantly between experiments and there is no generic documentation. GraphNeT developers are happy to collaborate with new experiments on this integration task; please contact us by opening a GitHub issue, if you are interested.”)
