### 2023-03-31 graphnet-team/graphnet#472 @asogaard
Improve structure and documentation of deployment functionality.
As pointed out [here](https://github.com/openjournals/joss-reviews/issues/4971#issuecomment-1487066422), the current version of the deployment functionality is entirely specific to IceCube, and it is not clear how it should be extended to other experiments.

I suggest:
* We introduce a minimal, generic `InferenceModule` class, that loads pretrained GraphNeT models, builds graphs, and applies the former to the latter to get predictions.
* Refactor `I3InferenceModule` to have it inherit from `InferenceModule` as one example of how to apply the same principles in an experiment-specific software stack.
* Write `SQLiteInferenceModule` and `ParquetInferenceModule` classes that inherit from `InferenceModule` and shows how to to sequential inference on generic data formats.
* Provide self-contained examples of how to apply the to above inference modules.
* Add a note in `graphnet.deployment` that we are happy to assist users wanting to extend this type of inference to other, experiment-specific data formats.

cc @RasmusOrsoe, @JostMigenda
