### 2023-03-31 #4971 @asogaard
Thanks for the update @JostMigenda — it's good to have you back. 😊  I think the above is a very reasonable point, and a pretty manageable one. I have opened this issue in `graphnet`: https://github.com/graphnet-team/graphnet/issues/472 to allow us to discuss potential solutions and work on these. Once we reach an acceptable solution to this, do I take it that you find the `graphnet` package and companion paper to be acceptable for JOSS publication?
