### 2023-04-18 #4971 @asogaard
@GageDeZoort, I just wanted to highlight that we have added a [`GETTING_STARTED.md`](https://github.com/graphnet-team/graphnet/blob/main/GETTING_STARTED.md) file at the root of the repository. This is exactly aimed at being _"a write-up (e.g. text surrounding Jupyter notebook cells) showing an example workflow you might hand to a student looking to develop their own GNN"_ which you requested early on. Please let me know how it looks to you, and if you have any other things that you consider to be acceptance-blockers.
