### 2023-04-18 #4971 @GageDeZoort
Thanks for the update - this looks very nice, I'm glad to see the mention of users providing more GNN functionality. With this and the answers you provided above considered, I'm satisfied and have checked all the review boxes!
