### 2023-04-18 #4971 @JostMigenda
The `GETTING_STARTED.md` looks very nice indeed! I see there are still a few `@TODO` markers left, so I hope you will be able to fill in those sections soon. However, that should not delay this review any longer.

@dfm I am very happy to recommend that JOSS **accept** this submission! 🎉
