### 2023-04-19 #4971 @dfm
@asogaard — I'm going through the final checks here, and I may have some small edits to the paper which I'll provide as a PR. In the meantime, can you:

1. Take one last read through the manuscript (linked by the bot above ☝️) to make sure that you're happy with it (it's harder to make changes later!), especially the author names and affiliations.
3. Create an archived release of that version of the software (using Zenodo or something similar). _Please make sure that the metadata (title and author list) exactly match the JOSS paper._ Then report the DOI of the release back to this thread.
