### 2023-04-19 #4971 @asogaard
Hi @dfm,

Thanks for getting us to the finish line! 🚀  I will read through the manuscript again and report back here with a DOI for the archived release of the software corresponding to the paper. Related to that: Should I just report the incremented version number here in the thread, or do I need to enter it somewhere?
