### 2023-04-19 #4971 @JostMigenda
Huh. Wasn't there some command for the editorialbot to set the version? Or did it not list this because as a reviewer, I don't have permission to do that? 🤔
