### 2023-05-12 #4971 @asogaard
Hi @dfm,

Sorry for the wait, we just needed to get the last few loose ends in order.

1. I have added a missing line to the main figure in the paper and updated the author list to reflect contributors since the initial submission.
2. I have incremented the version number to **v1.0.0** (as of today)
3. The Zenodo DOI corresponding to this version is **10.5281/zenodo.7928487**. I have checked that the metadata match the paper draft exactly.

Please let me know if you need anything else from me. :)
